### 2019-03-15 #1323 @sambitdash
From the said site I gather JOSS is interested in the following areas:

> JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; **extracts knowledge from large data sets**; offers a mathematical library, or similar.

There is no in-built native tool in Julia to read PDF files and extract text from documents. This tool enables you to achieve that. I guess from that perspective, it fits into the area of interest JOSS has defined for itself. Implementation of PDF Library from scratch is a fairly non-trivial task which has been achieved in this exercise using a new programming language.
