### 2019-03-15 #1323 @sambitdash
@arfon Sorry I may have misunderstood. Are you suggesting, I provide a detailed description as `paper.md` template? Should I check it in the PDFIO.jl repositiory?
