### 2019-03-15 #1323 @arfon
> @arfon Sorry I may have misunderstood. Are you suggesting, I provide a detailed description as paper.md template? Should I check it in the PDFIO.jl repositiory?

Yes please. Please review the submission guidelines here: https://joss.readthedocs.io/en/latest/submitting.html#submitting-a-paper-to-joss
