### 2019-03-20 fa37e4a
Ensuring font cid gets populated for standard 14 fonts
Also ensures the Rectangle is mapped to Float32.
