### 2019-04-01 e7c02ae
Fixes content is split into multiple parts.
1. Content array can be an Indirect object
2. Array may contain indirect object references
