### 2019-04-01 8703255
Fix for www.stillhq.com/pdfdb not accessible any longer
