+++ test/runtests.jl
@@ -3,11 +3,34 @@ using PDFIO
 using PDFIO.PD
 using PDFIO.Cos
 using PDFIO.Common
+using ZipFile
 
 # Internal methods for testing only
 using PDFIO.Cos: parse_indirect_ref, decode_ascii85, CosXString, parse_value
 
 include("debugIO.jl")
+
+pdftest_ver  = "0.0.1"
+pdftest_link = "https://github.com/sambitdash/PDFTest/archive/v"*pdftest_ver
+
+zipfile = "pdftest-"*pdftest_ver
+pdftest_link *= ".zip"
+zipfile *= ".zip"
+
+isfile(zipfile) || download(pdftest_link, zipfile)
+r = ZipFile.Reader(zipfile)
+buf = Vector{UInt8}(undef, 64*1024)
+for f in r.files
+    println("Filename: $(f.name)")
+    if f.method == ZipFile.Store
+        isdir(f.name) || mkdir(f.name)
+    elseif f.method == ZipFile.Deflate
+        isfile(f.name) ||
+            write(f.name, read(f, Vector{UInt8}(undef, f.uncompressedsize)))
+    end
+end
+close(r)
+pdftest_dir="PDFTest-"*pdftest_ver*"/"
 
 @testset "PDFIO tests" begin
     @testset "Miscellaneous" begin
@@ -145,10 +168,9 @@ include("debugIO.jl")
 
     @testset "General File Opening 3" begin
         @test begin
-            filename="3.pdf"
+            filename=pdftest_dir*"stillhq/3.pdf"
             DEBUG && println(filename)
-            isfile(filename)||
-                download("http://www.stillhq.com/pdfdb/000003/data.pdf",filename)
+            @assert isfile(filename)
             doc = pdDocOpen(filename)
             @assert pdDocGetPageCount(doc) == 30
             page = pdDocGetPage(doc, 1)
@@ -218,10 +240,9 @@ include("debugIO.jl")
 
     @testset "Test RunLengthDecode" begin
         @test begin
-            filename="582.pdf"
+            filename=pdftest_dir*"stillhq/582.pdf"
             DEBUG && println(filename)
-            isfile(filename)||
-                download("http://www.stillhq.com/pdfdb/000582/data.pdf",filename)
+            @assert isfile(filename)
             doc = pdDocOpen(filename)
             info = pdDocGetInfo(doc)
             @assert info["Trapped"] == cn"False"
@@ -239,47 +260,44 @@ include("debugIO.jl")
 
     @testset "Test ASCIIHexDecode" begin
         @test begin
-        filename="325.pdf"
-        DEBUG && println(filename)
-        isfile(filename)||
-            download("http://www.stillhq.com/pdfdb/000325/data.pdf",filename)
-        doc = pdDocOpen(filename)
-        @assert pdDocGetPageCount(doc) == 1
-        obj = cosDocGetObject(doc.cosDoc, CosIndirectObjectRef(7, 0))
-        stm=get(obj)
-        data=read(stm)
-        close(stm)
-        @assert length(data) == 121203
-        pdDocClose(doc)
-        length(utilPrintOpenFiles()) == 0
+            filename=pdftest_dir*"stillhq/325.pdf"
+            DEBUG && println(filename)
+            @assert isfile(filename)
+            doc = pdDocOpen(filename)
+            @assert pdDocGetPageCount(doc) == 1
+            obj = cosDocGetObject(doc.cosDoc, CosIndirectObjectRef(7, 0))
+            stm=get(obj)
+            data=read(stm)
+            close(stm)
+            @assert length(data) == 121203
+            pdDocClose(doc)
[…]
