+++ Project.toml
@@ -22,6 +22,7 @@ julia = "1.0"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
 
 [targets]
-test = ["Test"]
+test=["Test", "ZipFile"]
