### 2019-04-02 5dd35cb
More granular type mapping for better dispatch
