+++ src/PDDocImpl.jl
@@ -5,7 +5,7 @@ using ..Cos: CosDocImpl
 mutable struct PDDocImpl <: PDDoc
     cosDoc::CosDocImpl
     catalog::CosIndirectObject{CosDict}
-    pages::CosObject
+    pages::IDDN{CosDict}
     structTreeRoot::CosObject
     isTagged::Symbol #Valid values :tagged, :none and :suspect
     fonts::Dict{CosObject, PDFont}
@@ -83,7 +83,7 @@ seemingly non-standard way of computing page count. However, PDF spec does not
 discount the possibility of an intermediate node having page and pages nodes
 in the kids array. Hence, this implementation.
 =#
-function find_page_from_treenode(node::CosObject, pageno::Int)::CosObject
+function find_page_from_treenode(node::IDD{CosDict}, pageno::Int)
     mytype = get(node, cn"Type")
     #If this is a page object the pageno has to be 1
     if mytype == cn"Page"
@@ -96,7 +96,7 @@ function find_page_from_treenode(node::CosObject, pageno::Int)::CosObject
     for kid in kidsarr
         cnt = Cos.get_internal_pagecount(kid)
         (sum + cnt) >= pageno &&
-            return find_page_from_treenode(kid, pageno-sum)::CosObject
+            return find_page_from_treenode(kid, pageno-sum)
         sum += cnt
     end
     throw(ErrorException(E_INVALID_PAGE_NUMBER))
@@ -121,7 +121,7 @@ function update_structure_tree!(doc::PDDocImpl)
     return nothing
 end
 
-get_pd_font!(doc::PDDocImpl, cosfont::CosObject) =
+get_pd_font!(doc::PDDocImpl, cosfont::IDD{CosDict}) =
     get!(doc.fonts, cosfont, PDFont(doc, cosfont))
 
 get_pd_xobject!(doc::PDDocImpl, cosxobj::CosObject) = 
