+++ src/PDFonts.jl
@@ -52,7 +52,7 @@ mutable struct FontUnicodeMapping
 end
 
 function merge_encoding!(fum::FontUnicodeMapping, encoding::CosName,
-                         doc::CosDoc, font::CosObject)
+                         doc::CosDoc, font::IDDRef{CosDict})
     encoding_mapping =
         encoding == cn"WinAnsiEncoding"   ? WINEncoding_to_Unicode :
         encoding == cn"MacRomanEncoding"  ? MACEncoding_to_Unicode :
@@ -69,7 +69,7 @@ end
 # supported.
 # Font subset is addressed with font name identification.
 function merge_encoding!(fum::FontUnicodeMapping, encoding::CosNullType,
-                        doc::CosDoc, font::CosObject)
+                        doc::CosDoc, font::IDDRef{CosDict})
     subtype  = cosDocGetObject(doc, font, cn"Subtype")
     subtype !== cn"Type1" && subtype !== cn"MMType1" && return fum
     basefont = cosDocGetObject(doc, font, cn"BaseFont")
@@ -83,8 +83,8 @@ function merge_encoding!(fum::FontUnicodeMapping, encoding::CosNullType,
 end
 
 function merge_encoding!(fum::FontUnicodeMapping,
-                         encoding::Union{CosDict, CosIndirectObject{CosDict}},
-                         doc::CosDoc, font::CosObject)
+                         encoding::IDD{CosDict},
+                         doc::CosDoc, font::IDDRef{CosDict})
     baseenc = cosDocGetObject(doc, encoding, cn"BaseEncoding")
     merge_encoding!(fum, baseenc, doc, font)
     # Add the Differences
@@ -107,7 +107,8 @@ function merge_encoding!(fum::FontUnicodeMapping,
     return fum
 end
 
-function merge_encoding!(fum::FontUnicodeMapping, doc::CosDoc, font::CosObject)
+function merge_encoding!(fum::FontUnicodeMapping, doc::CosDoc,
+                         font::IDDRef{CosDict})
     encoding = cosDocGetObject(doc, font, cn"Encoding")
     merge_encoding!(fum, encoding, doc, font)
     toUnicode = cosDocGetObject(doc, font, cn"ToUnicode")
@@ -117,7 +118,7 @@ end
 
 function merge_encoding!(fum::FontUnicodeMapping,
                          cmap::CosIndirectObject{CosStream},
-                         doc::CosDoc, font::CosObject)
+                         doc::CosDoc, font::IDDRef{CosDict})
     stm_cmap = get(cmap)
     try
         fum.cmap = read_cmap(stm_cmap)
@@ -143,7 +144,7 @@ function update_glyph_id_std_14(cosfont, glyph_name_to_cid, cid_to_glyph_name)
     return true
 end
 
-function get_glyph_id_mapping(cosdoc::CosDoc, cosfont::CosObject)
+function get_glyph_id_mapping(cosdoc::CosDoc, cosfont::IDD{CosDict})
     glyph_name_to_cid, cid_to_glyph_name =
         Dict{CosName, UInt8}(), Dict{UInt8, CosName}()
     cosfont === CosNull && return glyph_name_to_cid, cid_to_glyph_name
@@ -201,7 +202,7 @@ get_encoded_string(s::CosString, fum::FontUnicodeMapping) =
     return String(NativeEncodingToUnicode(v, fum.encoding))
 end
 
-function get_unicode_chars(b::UInt8, i::Interval, v::CosObject)
[…]
