+++ src/PDXObject.jl
@@ -29,16 +29,18 @@ end
 
 mutable struct PDDefaultXObject <: PDXObject
     doc::PDDoc
-    obj::CosObject
+    obj::IDD{CosDict}
 end
 
-function createPDXObject(doc::PDDoc, cosstm::CosObject)
-    otype = get(cosstm, cn"Type")
+function createPDXObject(doc::PDDoc,
+                         cossd::Union{CosIndirectObject{CosStream},
+                                      IDD{CosDict}})
+    otype = get(cossd, cn"Type")
     @assert otype === cn"XObject" || otype === CosNull
-    subtype = get(cosstm, cn"Subtype")
-    subtype === cn"Form"  && return PDFormXObject(doc, cosstm)
-    subtype === cn"Image" && return PDImageXObject(doc, cosstm)
-    return PDDefaultXObject(doc, cosobj)
+    subtype = get(cossd, cn"Subtype")
+    subtype === cn"Form"  && return PDFormXObject(doc, cossd)
+    subtype === cn"Image" && return PDImageXObject(doc, cossd)
+    return PDDefaultXObject(doc, cossd)
 end
 
 function find_resource(xobj::PDFormXObject,
