### 2019-04-07 454bc12
Fixes the following:
1. More type related issues.
2. The content operator RG being misdiagnosed as CosIndirectObjectRef.
