+++ src/PDPageElement.jl
@@ -521,11 +521,13 @@ mutable struct GState{T}
     GState{T}() where T = new(init_graphics_state())
 end
 
-new_gstate(state::GState{T}) where {T} = GState{T}()
-
-Base.setindex!(gs::GState, v::Any, k::Symbol) = (gs.state[end][k] = v)
-Base.getindex(gs::GState, k::Symbol) = gs.state[end][k]
-Base.get(gs::GState, k::Symbol, defval::Any) = get(gs.state[end], k, defval)
+new_gstate(state::GState{T}) where T = GState{T}()
+Base.setindex!(gs::GState, v::V, k::Symbol) where V = (gs.state[end][k] = v)
+Base.get(gs::GState, k::Symbol, R::Type) = gs.state[end][k]::R
+Base.get(gs::GState, k::Symbol, defval::S, R::Type=S) where S  =
+    get(gs.state[end], k, defval)::Union{R, S}
+Base.get!(gs::GState, k::Symbol, defval::S, R::Type=S) where S =
+    get!(gs.state[end], k, defval)::Union{R, S}
 Base.delete!(gs::GState, k::Symbol) = delete!(gs.state[end], k)
 save!(gs::GState) = (push!(gs.state, copy(gs.state[end])); gs)
 restore!(gs::GState) = (pop!(gs.state); gs)
@@ -554,9 +556,9 @@ end
 
 function show_text_layout!(io::IO, state::GState)
     #Make sure to deepcopy. Otherwise the data structures will be lost
-    heap::Vector{TextLayout} = deepcopy(state[:text_layout])
+    heap = deepcopy(get(state, :text_layout, Vector{TextLayout}))
     sort!(heap, lt= > )
-    szdict::Dict{Int, Int} = state[:h_profile]
+    szdict = get(state, :h_profile, Dict{Int, Int})
 
     x = 0f0
     y = -1f0
@@ -626,21 +628,23 @@ end
 
 @inline function evalContent!(tr::PDPageTextRun, state::GState)
     evalContent!(tr.elem, state)
-    tfs::Float32 = get(state, :fontsize, 0f0)
-    th::Float32  = state[:Tz]/100f0
-    ts::Float32  = state[:Ts]
-    tc::Float32  = state[:Tc]
-    tw::Float32  = state[:Tw]
-    tm::Matrix{Float32}  = state[:Tm]
-    ctm::Matrix{Float32} = state[:CTM]
+    tfs = get(state, :fontsize, 0f0)
+    th  = get(state, :Tz, Float32)/100f0
+    ts  = get(state, :Ts, Float32)
+    tc  = get(state, :Tc, Float32)
+    tw  = get(state, :Tw, Float32)
+    tm  = get(state, :Tm, Matrix{Float32})
+    ctm = get(state, :CTM, Matrix{Float32})
     trm = tm*ctm
 
-    fontname, font = get(state, :font, (cn"", CosNull))
+    (fontname, font) = get(state, :font,
+                           (cn"", CosNull),
+                           Tuple{CosName, PDFont})
 
-    heap::Vector{TextLayout} = state[:text_layout]
+    heap = get(state, :text_layout, Vector{TextLayout})
     text, w, h = get_TextBox(tr.ss, font, tfs, tc, tw, th)
 
-    d::Dict{Int, Int} = state[:h_profile]
+    d = get(state, :h_profile, Dict{Int, Int})
     ih = round(Int, h*10)
     d[ih] = get(d, ih, 0) + length(text)
 
@@ -692,7 +696,7 @@ evalContent!(pdo::PDPageElement{:Q}, state::GState) = restore!(state)
     e = get(pdo.operands[5])
     f = get(pdo.operands[6])
[…]
