### 2019-04-13 9a89df8
Adding type fixes.
