### 2019-04-15 7c762cd
Implementation for pdDocGetOutline()
