+++ test/runtests.jl
@@ -4,8 +4,7 @@ using PDFIO.PD
 using PDFIO.Cos
 using PDFIO.Common
 using ZipFile
-
-@test [] == detect_ambiguities(Base, Core, PDFIO)
+using AbstractTrees
 
 # Internal methods for testing only
 using PDFIO.Cos: parse_indirect_ref, decode_ascii85, CosXString, parse_value
@@ -129,129 +128,6 @@ pdftest_dir="PDFTest-"*pdftest_ver*"/"
             pdDocClose(doc)
             length(utilPrintOpenFiles()) == 0
         end
-    end
-
-    @testset "Outlines" begin
-        @test begin
-            # This file has label catalog and usees /Dest outline entry (12.3.3 Document Outline)
-            filename="431.pdf"
-            DEBUG && println(filename)
-            isfile(filename) ||
-                download("http://www.stillhq.com/pdfdb/000431/data.pdf",filename)
-            doc = pdDocOpen(filename)
-            outline = pdDocGetOutline(doc)
-            @assert outline !== nothing
-            @assert length(outline) >= 20
-            @assert outline[1].Title == outline[1][:Title] == "Table of Contents"
-            @assert string(outline[1].PageRef) == "449 0 R"
-            @assert outline[1].PageNo == 2
-            @assert outline[1].Level == 1
-            @assert string(outline[19:20]) == string("""
-11. Advanced Topics - Linux Boot Process | PageRef=>509 0 R | PageNo=>32 | Expanded=>false
-  11.1. References for Boot Process | PageRef=>513 0 R | PageNo=>34""")
-            pdDocClose(doc)
-            length(utilPrintOpenFiles()) == 0
-        end
-        # This file created by LaTeX usees /A outline entry (12.3.3 Document Outline)
-        filename="files/outline.pdf"
-        DEBUG && println(filename)
-        doc = pdDocOpen(filename)
-        @test begin
-            outline = pdDocGetOutline(doc, add_index = true)
-            @assert outline !== nothing
-            @assert string(outline[1:2]) == """
-[1]: Chapter AA! | PageRef=>102 0 R | PageNo=>4 | Expanded=>false
-[2][1]: Section AA.aa! | PageRef=>106 0 R | PageNo=>5 | Expanded=>false
-[2][2][1]: SubSection AA.aa.a! | PageRef=>106 0 R | PageNo=>5 | Expanded=>false
-[2][2][2][1]: SubSubSection AA.aa.a.a! | PageRef=>106 0 R | PageNo=>5 | Expanded=>false
-[2][2][2][2][1]: Paragraph AA.aa.a.a.a! | PageRef=>106 0 R | PageNo=>5 | Expanded=>false
-[2][2][2][2][2][1]: SubParagraph AA.aa.a.a.a.a! | PageRef=>106 0 R | PageNo=>5
-[2][3]: Section AA.bb! | PageRef=>106 0 R | PageNo=>5 | Expanded=>false
-[2][4][1]: SubSection AA.bb.a! | PageRef=>106 0 R | PageNo=>5 | Expanded=>false
-[2][4][2][1]: SubSubSection AA.bb.a.a! | PageRef=>110 0 R | PageNo=>6 | Expanded=>false
-[2][4][2][2][1]: Paragraph AA.bb.a.a.a! | PageRef=>110 0 R | PageNo=>6 | Expanded=>false
-[2][4][2][2][2][1]: SubParagraph AA.bb.a.a.a.a! | PageRef=>110 0 R | PageNo=>6
[…]
