+++ src/PDOutline.jl
@@ -1,13 +1,11 @@
 export PDOutline,
-        items_count,
-        items,
-    PDOutlineItem,
-        hasproperty,
-    PDOutlineIter
+    PDOutlineItem
 
-using LabelNumerals
-using RomanNumerals
 using ..Cos
+using AbstractTrees
+import AbstractTrees: children, printnode
+
+abstract type _ParentNode  end
 
 ####################################################################################
 ## Data Structures
@@ -17,50 +15,67 @@ using ..Cos
 ```
 Representation of PDF document Outline item.
 
-It is currently defined as subtype of `AbstractDict{Symbol, Any}`.
-Check description of function `pdDocGetOutline` for more information.
-
-Methods which operate on this structure:
-- `propertynames(::PDOutlineItem; private = false) -> Set{Symbol}` - return properties present in this item.
-- `hasproperty(::PDOutlineItem, s::Symbol) -> Bool` - return weather item has given property.
 """
-struct PDOutlineItem <: AbstractDict{Symbol, Any}
-    props::Dict{Symbol, Any}
-    function PDOutlineItem()
-        new(Dict{Symbol, Any}())
-    end
-    function PDOutlineItem(p::Dict{Symbol, Any})
-        new(p)
-    end
+mutable struct PDOutlineItem <: _ParentNode
+    doc::PDDoc
+    cosdict::ID{CosDict}
+    parent::_ParentNode
+    prev::Union{PDOutlineItem, Nothing}
+    next::Union{PDOutlineItem, Nothing}
+    first::Union{PDOutlineItem, Nothing}
+    last::Union{PDOutlineItem, Nothing}
+    PDOutlineItem(doc::PDDoc, cosdict::ID{CosDict}, parent::_ParentNode) = 
+        new(doc, cosdict, parent, nothing, nothing, nothing, nothing)
 end
-Base.length(i::PDOutlineItem) = length(getfield(i, :props))
-Base.keys(i::PDOutlineItem) = keys(getfield(i, :props))
-Base.get(i::PDOutlineItem, k::Symbol, def) = get(getfield(i, :props), k, def)
-Base.get(i::PDOutlineItem, k::Integer, def) = throw(KeyError(string(k)))
-Base.iterate(i::PDOutlineItem) = iterate(getfield(i, :props))
-Base.iterate(i::PDOutlineItem, s::Int) = iterate(getfield(i, :props), s)
 
-function Base.getproperty(i::PDOutlineItem, s::Symbol)
-    try
-        s == :props && return getfield(i, :props)
-        if s == :Level
-            haskey(getfield(i, :props), :Level) && return getfield(i, :props)[:Level]
-            haskey(getfield(i, :props), :Index) && return length(getfield(i, :props)[:Index])
-            return 1
-        else
-            return getfield(i, :props)[s]
-        end
-    catch KeyError
-        return missing
+show(io::IO, x::PDOutlineItem) = showref(io, x.cosdict)
+
+populate_outline_items(parent::_ParentNode,
+                       ::CosNullType, ::CosNullType) = (nothing, nothing)
+
+@inline function populate_outline_items(parent::_ParentNode,
+                                        first_obj::ID{CosDict},
+                                        last_obj::ID{CosDict})
+    cosdoc = parent.doc.cosDoc
+    curr_obj = first_obj
+    prev_obj = cosDocGetObject(cosdoc, curr_obj, cn"Prev")
+    @assert prev_obj === CosNull
+        "Outline first item with invalid /Prev attribute"
+    last = first = curr = PDOutlineItem(parent.doc, curr_obj, parent)
+    curr.prev = nothing
+    while curr_obj !== last_obj
[…]
