+++ test/debugIO.jl
@@ -6,6 +6,7 @@ DEBUG=true. default is false.
 This is kept external to the regular library to ensure normal library operations
 are not affected.
 =#
+using Printf
 
 const DEBUG=false
 
@@ -28,7 +29,7 @@ end
 function util_open(filename, mode)
     global IODebug
     io=open(filename, mode)
-    @printf "Opening file: %s\n" filename
+    println("Opening file: ", filename)
     push!(IODebug[2], (filename, io))
     return io
 end
@@ -38,7 +39,7 @@ function util_close(handle::IOStream)
     idx=1
     for file in IODebug[2]
         if (handle === file[2])
-            @printf "Closing file: %s, %d\n" file[1] idx
+            println("Closing file: ", file[1], " ", idx)
             close(handle)
             deleteat!(IODebug[2],idx)
             return
@@ -80,15 +81,4 @@ function extract_text(io, doc)
             println(io)
         end
     end
-end
-
-function testfiles(filename)
-    name, ext = splitext(filename)
-    return (name*".res", joinpath(@__DIR__, "files", name*".txt"))
-end
-
-function local_testfiles(filename)
-    name, ext = splitext(filename)
-    return (name*".res", joinpath(@__DIR__, "files", name*".txt"),
-            joinpath(@__DIR__, "files", filename))
 end
