+++ test/files/outline.tex
@@ -1,128 +1,0 @@
-\documentclass[a4paper,10pt]{report}
-\usepackage[utf8]{inputenc}
-\usepackage{hyperref}
-\hypersetup{bookmarksdepth=subparagraph}
-% \hypersetup{bookmarksopenlevel=1}
-
-%opening
-\title{Outline}
-\author{Grzegorz Wierzchowski}
-
-\begin{document}
-
-\maketitle
-
-\begin{abstract}
-This is special document crafted to test \verb|pdDocGetOutline()| API function of \verb|PDFIO| Julia language library.
-\end{abstract}
-
-\tableofcontents
-
-% --------------------------
-\chapter{Chapter AA!}
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam quis nulla turpis. Sed sollicitudin
-neque vel ante laoreet, at aliquam sem fringilla. Ut ex metus, vestibulum quis porttitor et, volutpat
-vitae est. Cras in ornare arcu. Ut congue sem est, vitae pulvinar justo commodo et. Morbi cursus
-velit eget diam pellentesque dignissim. Sed feugiat neque vel augue fermentum porttitor. Mauris
-interdum velit sit amet vehicula tincidunt.
-Vestibulum gravida vulputate dolor, et scelerisque orci consectetur id. Ut congue ultricies dui eget
-dignissim. Nunc accumsan tempus viverra. Nullam posuere porttitor posuere. Donec finibus luctus
-quam eu mollis. Nulla pretium urna eu maximus dapibus. Quisque ac dui eu neque pretium
-accumsan id dictum nunc. Aliquam gravida pulvinar lorem. Proin vitae gravida nibh.
-Vivamus malesuada sollicitudin quam, nec malesuada libero. Nulla et laoreet urna. Donec justo
-ante, suscipit a dui sit amet, fermentum rhoncus arcu. Quisque aliquet lacus vitae rhoncus
-consectetur. Mauris tempor lacus ut consectetur ultricies. Duis id luctus mi. Nunc vitae hendrerit
-lacus. Proin at sapien leo.
-Suspendisse vitae sem blandit, auctor nisl in, posuere risus. Cras vel sapien leo. Vestibulum luctus
-diam cursus nibh vestibulum, nec tincidunt est fermentum. Aliquam velit ante, tincidunt quis
-maximus sit amet, elementum nec tellus. Proin non molestie velit, a commodo massa. Aliquam
-fermentum cursus ante, eget consequat eros tristique eget. Curabitur egestas nibh vitae massa
-egestas, ut porttitor tellus dictum.
-Aenean hendrerit, lorem vitae aliquam ultricies, velit tortor venenatis ex, sed lobortis dui turpis non
-est. Proin ac ipsum sed nunc egestas placerat laoreet sed libero. In urna turpis, imperdiet sed nulla
-cursus, efficitur maximus ligula. Aenean metus nisi, interdum sed metus convallis, eleifend lacinia
-dolor. Ut congue blandit ante vitae elementum. In et lacus id arcu posuere porta vitae non nibh.
-Integer quis orci sed mauris consectetur varius quis ut est. Quisque dictum, nisi eget porta lacinia,
-ante erat finibus lacus, ut luctus dolor tortor sit amet ligula. Donec id gravida odio, non gravida
-tortor. Quisque dictum erat ut nisi consequat, quis ultrices neque tempus. Aliquam eu gravida nisl.
-Nunc erat ligula, convallis non nunc ut, varius luctus erat. Phasellus at dolor lobortis, mollis sapien
[…]
