+++ Project.toml
@@ -13,7 +13,6 @@ LabelNumerals = "2e475f56-4567-5c41-94ff-3f69a29378f9"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Rectangle = "9a9db56c-8f71-5460-add5-cb450131785e"
 RomanNumerals = "37834d88-8936-577c-80c9-1066ecf66832"
 WinRPM = "c17dfb99-b4f7-5aad-8812-456da1ad7187"
@@ -25,6 +24,7 @@ julia = "1.0"
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 
 [targets]
-test = ["Test", "ZipFile"]
+test = ["Test", "ZipFile", "Printf"]
