### 2019-04-15 e1b6a7c
PDOutline issues fixed. Test cases added
