+++ src/PDOutline.jl
@@ -1,5 +1,6 @@
 export PDOutline,
-    PDOutlineItem
+    PDOutlineItem,
+    pdOutlineItemGetAttr
 
 using ..Cos
 using AbstractTrees
@@ -7,8 +8,48 @@ import AbstractTrees: children, printnode
 
 abstract type _ParentNode  end
 
-####################################################################################
-## Data Structures
+"""
+```
+    PDDestination
+```
+"""
+
+struct PDDestination
+    pageno::Int
+    layout::CosName
+    values::Vector{Float32}
+    zoom::Float32
+end
+
+ifnull(x, y) = x !== CosNull ? x : y
+
+function PDDestination(doc::PDDoc, arr::CosArray)
+    v = get(arr)
+    page = pd_doc_get_pagenum(doc, v[1])
+    values = [0f0, 0f0, 0f0, 0f0]
+    zoom = 0f0
+    if     v[2] === cn"XYZ"
+        values[1] = ifnull(v[3], 0f0)
+        values[4] = ifnull(v[4], 0f0)
+        zoom      = ifnull(v[5], 0f0)
+    elseif v[2] === cn"Fit"
+    elseif v[2] === cn"FitH"
+        values[4] = ifnull(v[3], 0f0)
+    elseif v[2] === cn"FitV"
+        values[1] = ifnull(v[4], 0f0)
+    elseif v[2] === cn"FitR"
+        values[1] = ifnull(v[3], 0f0)
+        values[2] = ifnull(v[4], 0f0)
+        values[3] = ifnull(v[5], 0f0)
+        values[4] = ifnull(v[6], 0f0)
+    elseif v[2] === cn"FitB"
+    elseif v[2] === cn"FitBH"
+        values[4] = ifnull(v[3], 0f0)
+    elseif v[2] === cn"FitBV"
+        values[1] = ifnull(v[3], 0f0)
+    end
+    return PDDestination(page, v[2], values, zoom)
+end
 """
 ```
     PDOutlineItem
@@ -30,51 +71,106 @@ end
 
 show(io::IO, x::PDOutlineItem) = showref(io, x.cosdict)
 
-populate_outline_items(parent::_ParentNode,
-                       ::CosNullType, ::CosNullType) = (nothing, nothing)
+populate_outline_items!(parent::_ParentNode, ::CosNullType, ::CosNullType) =
+    (parent.first = parent.last = nothing)
 
-@inline function populate_outline_items(parent::_ParentNode,
-                                        first_obj::ID{CosDict},
-                                        last_obj::ID{CosDict})
+@inline function populate_outline_items!(parent::_ParentNode,
+                                         first_obj::ID{CosDict},
+                                         last_obj::ID{CosDict})
     cosdoc = parent.doc.cosDoc
     curr_obj = first_obj
     prev_obj = cosDocGetObject(cosdoc, curr_obj, cn"Prev")
     @assert prev_obj === CosNull
         "Outline first item with invalid /Prev attribute"
-    last = first = curr = PDOutlineItem(parent.doc, curr_obj, parent)
+    first = curr = PDOutlineItem(parent.doc, curr_obj, parent)
     curr.prev = nothing
-    while curr_obj !== last_obj
+    while true
+        cfirst_obj = cosDocGetObject(cosdoc, curr_obj, cn"First")
+        clast_obj  = cosDocGetObject(cosdoc, curr_obj, cn"Last")
+        populate_outline_items!(curr, cfirst_obj, clast_obj)
+
         next_obj = cosDocGetObject(cosdoc, curr_obj, cn"Next")
-        @assert next_obj !== CosNull
-            "Outline item with invalid /Next attribute"
[…]
