+++ Project.toml
@@ -17,6 +17,7 @@ Rectangle = "9a9db56c-8f71-5460-add5-cb450131785e"
 RomanNumerals = "37834d88-8936-577c-80c9-1066ecf66832"
 WinRPM = "c17dfb99-b4f7-5aad-8812-456da1ad7187"
 ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 
 [compat]
 julia = "1.0"
@@ -24,7 +25,6 @@ julia = "1.0"
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 
 [targets]
-test = ["Test", "ZipFile", "Printf"]
+test = ["Test", "ZipFile"]

[…]
