+++ src/PDFIO.jl
@@ -22,6 +22,7 @@ export  PDDoc,
             pdDocGetPage,
             pdDocGetPageCount,
             pdDocGetPageRange,
+            pdDocGetOutline,
         PDPage,
             pdPageGetContents,
             pdPageIsEmpty,
@@ -31,6 +32,7 @@ export  PDDoc,
             pdPageGetFonts,
             pdPageGetCropBox,
             pdPageExtractText,
+            pdPageGetPageNumber,
         PDPageObject,
             PDPageObjectGroup,
                 PDPage_BeginGroup, PDPage_EndGroup,
@@ -44,7 +46,10 @@ export  PDDoc,
             pdFontIsItalic,
             pdFontIsFixedW,
             pdFontIsAllCap,
-            pdFontIsSmallCap
+            pdFontIsSmallCap,
+        PDOutline,
+            PDOutlineItem,
+                pdOutlineItemGetAttr
 
 using .Cos
 export  CosDoc,
