### 2019-04-16 4c554b6
Refactoring and cleaning up unused code pieces.
