### 2019-04-16 bbcad13
Refactoring changes. (#50)
* pdDocGetOutline() implementation
Related to #35
Also fixes #44

* Removed PageLabel detail from outline; some more internal changes.

* Implementation for pdDocGetOutline()

* Moving test files to PDFTest repository.

* PDOutline issues fixed. Test cases added

* Refactoring and cleaning up unused code pieces.
