+++ src/PDFIO.jl
@@ -22,6 +22,7 @@ export  PDDoc,
             pdDocGetPage,
             pdDocGetPageCount,
             pdDocGetPageRange,
+            pdDocGetPageLabel,
             pdDocGetOutline,
         PDPage,
             pdPageGetContents,
