+++ docs/make.jl
@@ -3,7 +3,7 @@ push!(LOAD_PATH,"../src/")
 using Documenter, PDFIO
 
 makedocs(
-    format = :html,
+    format = Documenter.HTML(),
     sitename = "PDFIO",
     pages = [
         "intro.md",
@@ -15,7 +15,7 @@ deploydocs(
     repo   = "github.com/sambitdash/PDFIO.jl.git",
     target = "build",
     branch = "gh-pages",
-    julia  = "0.6",
+    julia  = "1.0",
     deps   = nothing,
     make   = nothing
 )
