+++ Project.toml
@@ -9,15 +9,16 @@ AdobeGlyphList = "715cd884-8f22-55db-b077-449d1f7eef73"
 BinDeps = "9e28174c-4ba2-5203-b857-d8d62c4213ee"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 LabelNumerals = "2e475f56-4567-5c41-94ff-3f69a29378f9"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Rectangle = "9a9db56c-8f71-5460-add5-cb450131785e"
 RomanNumerals = "37834d88-8936-577c-80c9-1066ecf66832"
 WinRPM = "c17dfb99-b4f7-5aad-8812-456da1ad7187"
 ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 
 [compat]
 julia = "1.0"
