### 2019-04-18 a70293e
Configuring travis for doc
