### 2019-04-18 834def2
Updating .travis for Documenter
