+++ .travis.yml
@@ -35,7 +35,7 @@ after_success:
   - julia -e 'import PDFIO; cd(joinpath(dirname(pathof(PDFIO)), "..")); using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
   # push documentation from gh-pages
   - julia -e 'using Pkg; Pkg.add("Documenter")'
-  - julia -e 'cd(dirname(pathof(PDFIO))); include(joinpath("docs", "make.jl"))'
+  - julia -e 'import PDFIO; cd(dirname(pathof(PDFIO))); include(joinpath("docs", "make.jl"))'
 deploy:
   provider: pages
   skip_cleanup: true
