+++ .travis.yml
@@ -34,8 +34,17 @@ after_success:
   # push coverage results to Codecov
   - julia -e 'import PDFIO; cd(joinpath(dirname(pathof(PDFIO)), "..")); using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
   # push documentation from gh-pages
-  - julia -e 'using Pkg; Pkg.add("Documenter")'
-  - julia -e 'import PDFIO; cd(joinpath(dirname(pathof(PDFIO)), "..")); include(joinpath("docs", "make.jl"))'
+  # - julia -e 'using Pkg; Pkg.add("Documenter")'
+  # - julia -e 'import PDFIO; cd(joinpath(dirname(pathof(PDFIO)), "..")); include(joinpath("docs", "make.jl"))'
+jobs:
+  include:
+    - stage: "Documentation"
+      julia: 1.0
+      os: linux
+      script:
+        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
+        - julia --project=docs/ docs/make.jl
+      after_success: skip  
 deploy:
   provider: pages
   skip_cleanup: true
