+++ .travis.yml
@@ -42,6 +42,7 @@ jobs:
       julia: 1.0
       os: linux
       script:
+        - julia -e 'using Pkg; Pkg.add("Documenter")'
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
         - julia --project=docs/ docs/make.jl
       after_success: skip  
