### 2019-04-18 fcf3fa9
Updating travis.yml for documenter
