+++ .travis.yml
@@ -1,48 +1,30 @@
-## Documentation: http://docs.travis-ci.com/user/languages/julia/
 language: julia
+
+julia:
+  - 1.0
+  - 1.1
+  - nightly
+
 os:
   - linux
   - osx
-julia:
-  - 1.0
-  - nightly
+
+matrix:
+ allow_failures:
+ - julia: nightly
+
 notifications:
   email: false
-git:
-  depth: 99999999
-env:
-  - DOCUMENTER_DEBUG=true
 
-## uncomment the following lines to allow failures on nightly julia
-## (tests will run but not make your overall status red)
-# matrix:
-#  allow_failures:
-#  - julia: nightly
-
-## uncomment and modify the following lines to manually install system packages
-#addons:
-#  apt: # apt-get for linux
-#    packages:
-#    - gfortran
-#before_script: # homebrew for mac
-#  - if [ $TRAVIS_OS_NAME = osx ]; then brew install gcc; fi
-
-## uncomment the following lines to override the default test script
-#script:
-#  - julia -e 'Pkg.clone(pwd()); Pkg.build("PDFIO"); Pkg.test("PDFIO"; coverage=true)'
 after_success:
-  # push coverage results to Codecov
   - julia -e 'import PDFIO; cd(joinpath(dirname(pathof(PDFIO)), "..")); using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
-  # push documentation from gh-pages
-  # - julia -e 'using Pkg; Pkg.add("Documenter")'
-  # - julia -e 'import PDFIO; cd(joinpath(dirname(pathof(PDFIO)), "..")); include(joinpath("docs", "make.jl"))'
+
 jobs:
   include:
     - stage: "Documentation"
       julia: 1.0
       os: linux
       script:
-        - julia -e 'using Pkg; Pkg.add("Documenter")'
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
         - julia --project=docs/ docs/make.jl
       after_success: skip  
