### 2019-04-18 8d27315
removing 1.0
