+++ .travis.yml
@@ -1,7 +1,6 @@
 language: julia
 
 julia:
-  - 1.0
   - 1.1
   - nightly
 
