### 2019-04-18 687479c
Adding travis config file for documentation
