+++ .travis.yml
@@ -25,6 +25,6 @@ jobs:
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
-        - julia --project=docs/ -e docs/make.jl
+        - julia --project=docs/ docs/make.jl
       after_success: skip  
  
