### 2019-04-18 64e9a53
Revert "Adding travis config file for documentation"
This reverts commit 687479ce31834b95f253a2069af72f5b671e11e3.
