### 2019-04-18 9c519cc
Modifying travis for doc
