### 2019-04-19 03e5120
Attempting auto-deployment of docs from travis
