### 2019-04-19 04ed957
Updated documentation and auto deployment
