+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"intro/#PDFIO-1","page":"PDFIO","title":"PDFIO","text":"","category":"section"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"PDFIO is a simple PDF API focused on processing text from PDF files. The APIs is kept fairly simple and follows the PDF-32000-1:2008 specification as downloaded from Adobe website. The updated ISO versions can be downloaded from ISO websites. The API does not consider those versions yet. The API is developed in Julia keeping in mind the need for the file format parsing in native form. More over the object structure of PDF makes the API extend-able.","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"Technically,the PDFIO does not fully fit into the stream based file access APIs like JuliaIO/FileIO. However, there is a general interest to submit this API in that direction to ensure it's considered as a native file format understood by Julia.","category":"page"},{"location":"intro/#About-PDF-1","page":"PDFIO","title":"About PDF","text":"","category":"section"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"For people new to PDF development, PDF is a highly cross referenced file format without any specific linear location to read content. You can consider this as a form of a tree which needs to be read or understood as reading nodes and branches of a   tree. Hence, if your application demands you need not have to scan the full file format to get your job done. This is very different thinking than most file formats that are parsed from beginning to end.","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"The second part to remember is PDF is a derivative of a Page Description Language (PDL) namely PostScript. The visual sanctity is considered most important than document structure. Although, the document structure was introduced in PDF as an afterthought a lot later, it's not core to the format. Unfortunately, the creators never focus too much on this as it's a post processing over a print stream.","category":"page"},{"location":"intro/#Intent-and-Limitations-1","page":"PDFIO","title":"Intent and Limitations","text":"","category":"section"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"The APIs are developed for text extraction as a major focus. The rendering and printing aspects of PDF are not provided enough thoughts. Secondly, a native script based language was chosen as PDF specification is highly misunderstood and interpretations significantly vary from document creator to creator. A script based language provides flexibility to fix issues as we discover more nuances in the new files we discover. Thirdly, every well-developed native library out in the market need connectors and PDF being a standard should have native support in a modern language. Although, one can claim PDF is not the most  ideal language for text processing, it's just ubiquitous and cannot be ignored.","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"Nothing stops anyone to extend this APIs into a fully developed PDF Library  that's available to both commercial as well as non-commercial licensing under a  flexible license model. I am happy to collaborate with anyone who sees value in   extending this library in that direction. MIT License looks most flexible to  the author for the time-being.","category":"page"},{"location":"intro/#Common-1","page":"PDFIO","title":"Common","text":"","category":"section"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"Error handling is currently fairly weak with only errors or asserts leading ","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO"
[…]
