+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"intro/#PDFIO-1","page":"PDFIO","title":"PDFIO","text":"","category":"section"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"(Image: *nix Status) (Image: Win status) (Image: codecov.io) (Image: ) (Image: )","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"PDFIO is a simple PDF API focused on processing text from PDF files. The APIs is kept fairly simple and follows the PDF-32000-1:2008 specification as downloaded from Adobe website. The updated ISO versions can be downloaded from ISO websites. The API does not consider those versions yet. The API is developed in Julia keeping in mind the need for the file format parsing in native form. More over the object structure of PDF makes the API extend-able.","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"Technically,the PDFIO does not fully fit into the stream based file access APIs like JuliaIO/FileIO. However, there is a general interest to submit this API in that direction to ensure it's considered as a native file format understood by Julia.","category":"page"},{"location":"intro/#About-PDF-1","page":"PDFIO","title":"About PDF","text":"","category":"section"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"For people new to PDF development, PDF is a highly cross referenced file format without any specific linear location to read content. You can consider this as a form of a tree which needs to be read or understood as reading a tree. Hence, if your application demands you need not have to scan the full file format to get your job done. This is very different thinking than most file formats that are parsed from beginning to end.","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"The second part to remember is PDF is a derivative of a Page Description Language (PDL) like PostScript. The visual sanctity is considered most important than document structure. Although, the document structure was introduced in PDF as an afterthought a lot later, it's not core to the format. Unfortunately, the creators never focus too much on this as it's a post processing over a print stream.","category":"page"},{"location":"intro/#Structure-of-API-1","page":"PDFIO","title":"Structure of API","text":"","category":"section"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"The API is segregated into 3 modules:","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"Common\nCos\nPD","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"Common module has general system access and file access and parsing APIs.","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"Cos module is the low level file format for PDF. Carousel Object Structure was original term proposed inside Adobe which later transformed into Acrobat. Cos layer has the object structure, definition and the cross references to access them.","category":"page"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"PD module is the higher level document access layer. Accessing PDF pages or extracting the content from there or understanding document rendering using fonts or image objects will be typically in this layer. Please note that many objects in the PD layer actually refer to the Cos structure. You can consider PD Layer as the business logic while Cos Layer as the database for it.","category":"page"},{"location":"intro/#Intent-and-Limitations-1","page":"PDFIO","title":"Intent and Limitations","text":"","category":"section"},{"location":"intro/#","page":"PDFIO","title":"PDFIO","text":"The APIs are developed for text extraction as a major focus. The rendering and printing a
[…]
