### 2019-04-25 8ecfd19
Performance improvements for pdPageExtractText
