+++ src/PDFonts.jl
@@ -193,7 +193,8 @@ get_encoded_string(s::CosString, fum::Nothing) = CDTextString(s)
 get_encoded_string(s::CosString, fum::FontUnicodeMapping) = 
     get_encoded_string(Vector{UInt8}(s), fum)
 
-@inline function get_encoded_string(v::Vector{UInt8}, fum::FontUnicodeMapping)
+@inline function get_encoded_string(v::Union{Vector{UInt8}, NTuple{N, UInt8}},
+                                    fum::FontUnicodeMapping) where N
     length(v) == 0 && return ""
     fum.hasCMap && return get_encoded_string(v, fum.cmap)
     return String(NativeEncodingToUnicode(v, fum.encoding))
@@ -253,7 +254,7 @@ end
 get_encoded_string(s::CosString, cmap::CMap) =
     get_encoded_string(Vector{UInt8}(s), cmap::CMap)
 
-function get_encoded_string(barr::Vector{UInt8}, cmap::CMap)
+function get_encoded_string(barr, cmap::CMap)
     cs = cmap.code_space
     rm = cmap.range_map
     l = length(barr)
@@ -273,16 +274,16 @@ function get_encoded_string(barr::Vector{UInt8}, cmap::CMap)
         # code space. So may need to decipher the existence of a single
         # byte vs 2-byte code from the range map. See `else` below.
         itree = xs[1][2]
-        # This case is very clearly a single byte range 
+        # This case is very clearly a single byte range
+        itv = intersect(rm, Interval(b1, b1))
         if itree === CosNull 
-            itv = intersect(rm, Interval(b1, b1))
             if length(itv) > 0
                 carr = get_unicode_chars(b1, itv[1][1], itv[1][2])
             else
                 push!(carr, Char(0))
             end
         else
-            itree1 = intersect(rm, Interval(b1, b1))
+            itree1 = itv 
             if length(itree1) == 0
                 push!(carr, Char(0))
                 continue
