### 2019-04-26 #1323 @sambitdash
@danielskatz submitted the paper in the Github folder. Kindly, review the submission and provide your feedback.
