### 2019-04-26 e7ead30
Fix the .bib file for paper submission.  (#61)
* Review comments to fix .bib file.
