+++ paper/paper.bib
@@ -1,9 +1,8 @@
 @book{Adobe:2008,
-  	Adsnote = {Provided by Adobe Systems Incorporarted},
-  	Adsurl = {https://www.adobe.com/devnet/pdf/pdf_reference.html},
-  	Author = {Adobe Systems Inc.},
-  	Booktitle = {Document Management - Portable Document Format - Part 1: PDF 1.7},
-  	Publisher = {Adobe Systems Incorporarted},
-  	Title = {{Document Management - Portable Document Format - Part 1: PDF 1.7}},
-  	Year = 2008
+  author = {{Adobe Systems Inc.}},
+  publisher = {{Adobe Systems Inc.}},
+  year = 2008,
+  title = {{Document Management - Portable Document Format - Part 1: PDF 1.7}},
+  url = {https://www.adobe.com/devnet/pdf/pdf_reference.html},
+  note = {Provided by the Adobe Systems Incorporarted}
 }
