### 2019-05-14 #1323 @alexhanna
@malmaud, @jameshclrk, @ChrisRackauckas -- are any of you available to edit this submission?
