### 2019-05-17 #1453 @whedon
[REVIEW]: Native PDF Reader Library in Julia
**Submitting author:** @sambitdash (<a href="http://orcid.org/0000-0003-4856-7244">Sambit Kumar Dash</a>)
**Repository:** <a href="https://github.com/sambitdash/PDFIO.jl" target ="_blank">https://github.com/sambitdash/PDFIO.jl</a>
**Version:** v0.1.8
