### 2019-05-17 sambitdash/PDFIO.jl#62 @malmaud
Tests fail
```
v1.1) pkg> test PDFIO
   Testing PDFIO
 Resolving package versions...
    Status `/var/folders/t6/ddh10c6j5r54sg19jlc59n580000gn/T/tmpCjMIM2/Manifest.toml`
  [1520ce14] AbstractTrees v0.2.1
  [715cd884] AdobeGlyphList v0.1.1
  [9e28174c] BinDeps v0.8.10
  [b99e7846] BinaryProvider v0.5.4
  [e1450e63] BufferedStreams v1.0.0
  [34da2185] Compat v2.1.0
  [ffbed154] DocStringExtensions v0.7.0
  [e30172f5] Documenter v0.22.4
  [0862f596] HTTPClient v0.2.1
  [682c06a0] JSON v0.20.0
  [2e475f56] LabelNumerals v0.1.0
  [b27032c2] LibCURL v0.5.0
  [522f3ed2] LibExpat v0.5.0
  [2ec943e9] Libz v1.0.0
  [4d0d745f] PDFIO v0.1.3
  [27ebfcd6] Primes v0.4.0
  [9a9db56c] Rectangle v0.1.1
  [37834d88] RomanNumerals v0.3.1
  [30578b45] URIParser v0.4.0
  [c17dfb99] WinRPM v0.4.2
  [a5390f91] ZipFile v0.8.1
  [2a0f44e3] Base64  [`@stdlib/Base64`]
  [ade2ca70] Dates  [`@stdlib/Dates`]
  [8bb1440f] DelimitedFiles  [`@stdlib/DelimitedFiles`]
  [8ba89e20] Distributed  [`@stdlib/Distributed`]
  [b77e0a4c] InteractiveUtils  [`@stdlib/InteractiveUtils`]
  [76f85450] LibGit2  [`@stdlib/LibGit2`]
  [8f399da3] Libdl  [`@stdlib/Libdl`]
  [37e2e46d] LinearAlgebra  [`@stdlib/LinearAlgebra`]
  [56ddb016] Logging  [`@stdlib/Logging`]
  [d6f4376e] Markdown  [`@stdlib/Markdown`]
  [a63ad114] Mmap  [`@stdlib/Mmap`]
  [44cfe95a] Pkg  [`@stdlib/Pkg`]
  [de0858da] Printf  [`@stdlib/Printf`]
  [3fa0cd96] REPL  [`@stdlib/REPL`]
  [9a3f8284] Random  [`@stdlib/Random`]
  [ea8e919c] SHA  [`@stdlib/SHA`]
  [9e88b42a] Serialization  [`@stdlib/Serialization`]
  [1a1011a3] SharedArrays  [`@stdlib/SharedArrays`]
  [6462fe0b] Sockets  [`@stdlib/Sockets`]
  [2f01184e] SparseArrays  [`@stdlib/SparseArrays`]
  [10745b16] Statistics  [`@stdlib/Statistics`]
  [8dfed614] Test  [`@stdlib/Test`]
  [cf7118a7] UUIDs  [`@stdlib/UUIDs`]
  [4ec0a83e] Unicode  [`@stdlib/Unicode`]
ERROR: LoadError: LoadError: could not open file /Users/malmaud/.julia/packages/ZipFile/YHTbb/deps/deps.jl
Stacktrace:
 [1] include at ./boot.jl:326 [inlined]
 [2] include_relative(::Module, ::String) at ./loading.jl:1038
 [3] include at ./sysimg.jl:29 [inlined]
 [4] include(::String) at /Users/malmaud/.julia/packages/ZipFile/YHTbb/src/Zlib.jl:26
 [5] top-level scope at none:0
 [6] include at ./boot.jl:326 [inlined]
 [7] include_relative(::Module, ::String) at ./loading.jl:1038
 [8] include at ./sysimg.jl:29 [inlined]
 [9] include(::String) at /Users/malmaud/.julia/packages/ZipFile/YHTbb/src/ZipFile.jl:36
 [10] top-level scope at none:0
 [11] include at ./boot.jl:326 [inlined]
 [12] include_relative(::Module, ::String) at ./loading.jl:1038
 [13] include(::Module, ::String) at ./sysimg.jl:29
 [14] top-level scope at none:2
 [15] eval at ./boot.jl:328 [inlined]
 [16] eval(::Expr) at ./client.jl:404
 [17] top-level scope at ./none:3
in expression starting at /Users/malmaud/.julia/packages/ZipFile/YHTbb/src/Zlib.jl:50
in expression starting at /Users/malmaud/.julia/packages/ZipFile/YHTbb/src/ZipFile.jl:43
ERROR: LoadError: Failed to precompile ZipFile [a5390f91-8eb1-5f08-bee0-b1d1ffed6cea] to /Users/malmaud/.julia/compiled/v1.1/ZipFile/cOum2.ji.
Stacktrace:
 [1] error(::String) at ./error.jl:33
 [2] compilecache(::Base.PkgId, ::String) at ./loading.jl:1197
 [3] _require(::Base.PkgId) at ./loading.jl:960
 [4] require(::Base.PkgId) at ./loading.jl:858
 [5] require(::Module, ::Symbol) at ./loading.jl:853
 [6] include at ./boot.jl:326 [inlined]
 [7] include_relative(::Module, ::String) at ./loading.jl:1038
 [8] include(::Module, ::String) at ./sysimg.jl:29
 [9] include(::String) at ./client.jl:403
 [10] top-level scope at none:0
in expression starting at /Users/malmaud/.julia/packages/PDFIO/28lLV/test/runtests.jl:6
ERROR: Package PDFIO errored during testing
```
