### 2019-05-17 #1453 @malmaud
I've begun my review.

In general, I think the need is clear - this package is a utility for reading PDF files in Julia, which is clearly useful to some people.

The paper is intelligible, although suffers from grammatical errors throughout that should be fixed (as does the documentation). An example of a particularly ungrammatical sentence is  "Almost, all the functionality of PDF understanding is entirely written from scratch in Julia with only exception of usage of certain (de)compression codecs".

I find the third paragraph ("The APIs are developed...") confusing. What is a "native script based language"? I also don't understand the meaning of the sentence "every well-developed native library out in the market need connectors".

In "Their contribution can be viewed from the following link.", the URL of the link should be included in the paper - I don't think a JOSS paper should have embedded hyperlinks without the actual URL being visible in the document.

The paper should reference other Julia packages for handling PDF documents and explain how this package differs. In particular, I believe `Taro.jl` is the dominant solution at the moment.

At present, I don't think the documentation is sufficient. There are no examples or tutorials referenced in the readme or documentation. The only overview, the "API structure and design" section in `https://sambitdash.github.io/PDFIO.jl/stable/#API-Structure-and-Design-1`, is minimal.

The documentation, as far as I've been able to find,  consists solely of a description of each of dozens of API method which has been autogenerated from the docstrings in the source code and is listed in alphabetical order. It is not reasonable to expect new users to read through that to get started, and I'm not personally willing to do so and so cannot personally test to see if this package is functional.

The package did install, although I can't personally test its functionality since the documentation is too sparse for me to understand how to use it.

Running the automated test (`Pkg.test("PDFIO")`) fails for me (see https://github.com/sambitdash/PDFIO.jl/issues/62). From the logs, it looks like the issue is ultimately with a different downstream package (`ZipFile`), but nevertheless I can't verify this package passes its own tests.

In conclusion, I think major revisions to the documentation are needed, but if that can be accomplished, this software is a good fit for JOSS.
