### 2019-05-18 #1453 @sambitdash
@malmaud Thanks for your review. I will update the documentation to address your concerns around the documentation. On the failing of the test cases, I guess for some reason the build step did not run when you added the package. If you can run the following command:

```
pkg> build PDFIO
```
the test should have passed successfully.
