### 2019-05-18 sambitdash/PDFIO.jl#62 @sambitdash
@malmaud Can you follow the following steps?

```
$ mkdir test
$ cd test
$ julia   # This should be Julia 1.0 or higher
v1.1 pkg> activate .
(test) pkg> add PDFIO
(test) pkg> test PDFIO
```
It seems for some reason `ZipFile.jl` may have failed on your system.

I am attaching my output for the command below:

```
(v1.1) pkg> activate .

(test) pkg> add PDFIO
  Updating registry at `~/.julia/registries/General`
  Updating git-repo `https://github.com/JuliaRegistries/General.git`
 Resolving package versions...
 Installed PDFIO ────── v0.1.3
 Installed Documenter ─ v0.22.4
  Updating `~/work/test/Project.toml`
  [4d0d745f] + PDFIO v0.1.3
  Updating `~/work/test/Manifest.toml`
  [1520ce14] + AbstractTrees v0.2.1
  [715cd884] + AdobeGlyphList v0.1.1
  [9e28174c] + BinDeps v0.8.10
  [b99e7846] + BinaryProvider v0.5.4
  [e1450e63] + BufferedStreams v1.0.0
  [34da2185] + Compat v2.1.0
  [ffbed154] + DocStringExtensions v0.7.0
  [e30172f5] + Documenter v0.22.4
  [0862f596] + HTTPClient v0.2.1
  [682c06a0] + JSON v0.20.0
  [2e475f56] + LabelNumerals v0.1.0
  [b27032c2] + LibCURL v0.5.0
  [522f3ed2] + LibExpat v0.5.0
  [2ec943e9] + Libz v1.0.0
  [4d0d745f] + PDFIO v0.1.3
  [27ebfcd6] + Primes v0.4.0
  [9a9db56c] + Rectangle v0.1.1
  [37834d88] + RomanNumerals v0.3.1
  [30578b45] + URIParser v0.4.0
  [c17dfb99] + WinRPM v0.4.2
  [a5390f91] + ZipFile v0.8.1
  [2a0f44e3] + Base64
  [ade2ca70] + Dates
  [8bb1440f] + DelimitedFiles
  [8ba89e20] + Distributed
  [b77e0a4c] + InteractiveUtils
  [76f85450] + LibGit2
  [8f399da3] + Libdl
  [37e2e46d] + LinearAlgebra
  [56ddb016] + Logging
  [d6f4376e] + Markdown
  [a63ad114] + Mmap
  [44cfe95a] + Pkg
  [de0858da] + Printf
  [3fa0cd96] + REPL
  [9a3f8284] + Random
  [ea8e919c] + SHA
  [9e88b42a] + Serialization
  [1a1011a3] + SharedArrays
  [6462fe0b] + Sockets
  [2f01184e] + SparseArrays
  [10745b16] + Statistics
  [8dfed614] + Test
  [cf7118a7] + UUIDs
  [4ec0a83e] + Unicode
  Building PDFIO → `~/.julia/packages/PDFIO/28lLV/deps/build.log`  <----- This line may have failed in your case

(test) pkg> test PDFIO
   Testing PDFIO
 Resolving package versions...
    Status `/tmp/tmpw5Rix0/Manifest.toml`
  [1520ce14] AbstractTrees v0.2.1
  [715cd884] AdobeGlyphList v0.1.1
  [9e28174c] BinDeps v0.8.10
  [b99e7846] BinaryProvider v0.5.4
  [e1450e63] BufferedStreams v1.0.0
  [34da2185] Compat v2.1.0
  [ffbed154] DocStringExtensions v0.7.0
  [e30172f5] Documenter v0.22.4
  [0862f596] HTTPClient v0.2.1
  [682c06a0] JSON v0.20.0
  [2e475f56] LabelNumerals v0.1.0
  [b27032c2] LibCURL v0.5.0
  [522f3ed2] LibExpat v0.5.0
  [2ec943e9] Libz v1.0.0
  [4d0d745f] PDFIO v0.1.3
  [27ebfcd6] Primes v0.4.0
  [9a9db56c] Rectangle v0.1.1
  [37834d88] RomanNumerals v0.3.1
  [30578b45] URIParser v0.4.0
  [c17dfb99] WinRPM v0.4.2
  [a5390f91] ZipFile v0.8.1
  [2a0f44e3] Base64  [`@stdlib/Base64`]
  [ade2ca70] Dates  [`@stdlib/Dates`]
  [8bb1440f] DelimitedFiles  [`@stdlib/DelimitedFiles`]
  [8ba89e20] Distributed  [`@stdlib/Distributed`]
  [b77e0a4c] InteractiveUtils  [`@stdlib/InteractiveUtils`]
  [76f85450] LibGit2  [`@stdlib/LibGit2`]
  [8f399da3] Libdl  [`@stdlib/Libdl`]
  [37e2e46d] LinearAlgebra  [`@stdlib/LinearAlgebra`]
  [56ddb016] Logging  [`@stdlib/Logging`]
  [d6f4376e] Markdown  [`@stdlib/Markdown`]
  [a63ad114] Mmap  [`@stdlib/Mmap`]
  [44cfe95a] Pkg  [`@stdlib/Pkg`]
  [de0858da] Printf  [`@stdlib/Printf`]
  [3fa0cd96] REPL  [`@stdlib/REPL`]
  [9a3f8284] Random  [`@stdlib/Random`]
  [ea8e919c] SHA  [`@stdlib/SHA`]
  [9e88b42a] Serialization  [`@stdlib/Serialization`]
  [1a1011a3] SharedArrays  [`@stdlib/SharedArrays`]
  [6462fe0b] Sockets  [`@stdlib/Sockets`]
  [2f01184e] SparseArrays  [`@stdlib/SparseArrays`]
  [10745b16] Statistics  [`@stdlib/Statistics`]
  [8dfed614] Test  [`@stdlib/Test`]
  [cf7118a7] UUIDs  [`@stdlib/UUIDs`]
  [4ec0a83e] Unicode  [`@stdlib/Unicode`]
                                 Dload  Upload   Total   Spent    Left  Speed
100   123    0   123    0     0     47      0 --:--:--  0:00:02 --:--:--    47
100 7314k    0 7314k    0     0   198k      0 --:--:--  0:00:36 --:--:-- 18737
Filename: PDFTest-0.0.3/
Filename: PDFTest-0.0.3/README.md
Filename: PDFTest-0.0.3/files/
Filename: PDFTest-0.0.3/files/1.pdf
Filename: PDFTest-0.0.3/files/16-969_o7jp.pdf
Filename: PDFTest-0.0.3/files/16.cmap
Filename: PDFTest-0.0.3/files/1_noinfo.pdf
Filename: PDFTest-0.0.3/files/8.cmap
Filename: PDFTest-0.0.3/files/A1947-14.pdf
Filename: PDFTest-0.0.3/files/A1947-15.pdf
Filename: PDFTest-0.0.3/files/Pratham-Sanskaran.pdf
Filename: PDFTest-0.0.3/files/empty_property.pdf
Filename: PDFTest-0.0.3/files/fileAttachment.pdf
Filename: PDFTest-0.0.3/files/fileAttachment_fileAttachmentAnnotation.pdf
Filename: PDFTest-0.0.3/files/outline.pdf
Filename: PDFTest-0.0.3/files/pdf-17.pdf
Filename: PDFTest-0.0.3/files/pdf-sample.pdf
Filename: PDFTest-0.0.3/files/spec-2.pdf
Filename: PDFTest-0.0.3/files/spec-2c.pdf
Filename: PDFTest-0.0.3/source/
Filename: PDFTest-0.0.3/source/outline.tex
Filename: PDFTest-0.0.3/stillhq/
Filename: PDFTest-0.0.3/stillhq/3.pdf
Filename: PDFTest-0.0.3/stillhq/325.pdf
Filename: PDFTest-0.0.3/stillhq/339.pdf
Filename: PDFTest-0.0.3/stillhq/388.pdf
Filename: PDFTest-0.0.3/stillhq/431.pdf
Filename: PDFTest-0.0.3/stillhq/504.pdf
Filename: PDFTest-0.0.3/stillhq/582.pdf
Filename: PDFTest-0.0.3/stillhq/589.pdf
Filename: PDFTest-0.0.3/templates/
Filename: PDFTest-0.0.3/templates/16-969_o7jp.txt
Filename: PDFTest-0.0.3/templates/431.toc
Filename: PDFTest-0.0.3/templates/431.txt
Filename: PDFTest-0.0.3/templates/A1947-15.txt
Filename: PDFTest-0.0.3/templates/Graphics-wpf.txt
Filename: PDFTest-0.0.3/templates/Pratham-Sanskaran.txt
Filename: PDFTest-0.0.3/templates/outline.toc
Filename: PDFTest-0.0.3/templates/page5.txt
Filename: PDFTest-0.0.3/templates/spec-2.txt
Filename: PDFTest-0.0.3/templates/spec-2c.txt
                                 Dload  Upload   Total   Spent    Left  Speed
100   212    0   212    0     0     46      0 --:--:--  0:00:04 --:--:--    46
100   246  100   246    0     0     51      0  0:00:04  0:00:04 --:--:--     0
100  191k  100  191k    0     0  30087      0  0:00:06  0:00:06 --:--:--  226k
Test Summary: | Pass  Total
PDFIO tests   |   49     49
   Testing PDFIO tests passed
```
