### 2019-05-18 sambitdash/PDFIO.jl#62 @sambitdash
@malmaud or just try the following:

```
v 1.1 pkg> build PDFIO
```
