### 2019-05-21 sambitdash/PDFIO.jl#62 @malmaud
For some reason (I guess a bug in the package manager), that still failed with the same error. I just installed and tested in a new environment and the tests passed.
