### 2019-05-21 #1453 @malmaud
OK, I just installed and tested in a fresh Julia environment and the tests pass now.
