### 2019-06-09 #1453 @sambitdash
@labarba - I am going for a complete revamp of the documentation by the end of this month. Most of the suggested changes will be addressed as part of that. I will update when the task is accomplished.
