+++ docs/make.jl
@@ -1,3 +1,4 @@
+
 if Base.HOME_PROJECT[] !== nothing
     # JuliaLang/julia/pull/28625
     Base.HOME_PROJECT[] = abspath(Base.HOME_PROJECT[])
@@ -11,6 +12,7 @@ makedocs(
     pages = [
         "intro.md",
         "API Structure and Design" => "index.md"
+        "Digital Signatures" => "digsig.md"
     ]
 )
 
