+++ src/PDDoc.jl
@@ -10,7 +10,8 @@ export PDDoc,
        pdDocGetPageRange,
        pdDocHasPageLabels,
        pdDocGetPageLabel,
-       pdDocGetOutline
+       pdDocGetOutline,
+       pdDocValidateSignatures
 
 using ..Common
 
@@ -231,4 +232,147 @@ function pdDocGetOutline(doc::PDDoc)
     tocobj = cosDocGetObject(cosDoc, catalog, cn"Outlines")
     tocobj === nothing && return nothing
     return PDOutline(doc, tocobj)
+end
+
+"""
+```
+    pdDocIsSigned(doc::PDDoc) -> Bool
+```
+PDF documents can be signed and content can be approved by one or more reviewers.
+"""
+function pdDocIsSigned(doc::PDDoc)
+    catalog = pdDocGetCatalog(doc)
+    cosDoc = pdDocGetCosDoc(doc)
+    acroform = cosDocGetObject(cosDoc, catalog, cn"AcroForm")
+    acroform === CosNull && return false
+    sigfobj  = cosDocGetObject(cosDoc, acroform, cn"SigFlags")
+    sigfobj  === CosNull && return false
+    return get(sigfobj) & 0x1 != 0
+end
+
+"""
+```
+    pdDocValidateSignatures(doc::PDDoc; export_certs=false) -> Vector{Dict{Symbol, Any}}
+```
+## Input
+| param      | Description                                                  |
+|:-----------|:-------------------------------------------------------------|
+|doc         |The document for which all the signatures are to be validated.|
+|export_certs|Optional keyword parameter when set, exports all the          |
+|            |certificates that are embeded in the PDF document. These      |
+|            |certificates can be for end-entities or one or more certifying|
+|            |authorities.                                                  |
+|            |Certificates are exported to the file `<PDF filename>.pem`.   |
+
+## Output
+Vector of dictionary objects representing one dictionary object for each
+signature. The dictionary objects map the symbols to output as per the following
+table. 
+
+|Symbol    |Description                                                  |
+|:---------|:------------------------------------------------------------|
+|:Name     |The name of the person or authority signing the document.    |
+|:P        |Object reference of the page in which the signature is found.|     
+|:M        |The `CDDate` when the document was signed.                   |     
+|:certs    |The certificates associated with every signature object.     |     
+|:subfilter|The subfilter of PDF signature object.                       |
+|:FQT      |Fully qualified title of the signature form.                 |
+|:chain    |The certificate chain that validated the signature.          |
+|:passed   |Validation status of the signature (true / false)            |
+|:error_message| Error message returned during the validation            |
+|:stacktrace| The stack dump of where the validation failure occurred    |
+
+## Notes
+1. Any additional certificates needed for validating a certificate trust chain
+   has to be added manually to the *Trust Store* file at:
+   `<Package Directory>/data/certs/cacerts.pem` in the PEM format. Normally,
+   certificate authorities (root as well as intermediate) are represented in the
+   trust store.
+2. Presence of an end-entity certificate in the *Trust Store* ensures that the
+   chain validation for the certificate does not have to be carried out. However,
+   this is not considered a good practice for certificates as the certificate
+   validation is an important attribute to avoid security breaches in the chain.
[…]
