### 2019-06-15 a8b6c63
Adding additional checks on test cases.
