### 2019-06-15 347468b
PDFIO requires zlib-1.2.8 or above or OpenSSL-1.1.0f or above
If the system library is not supporting that version the library will be
built from source.
