### 2019-06-16 d67e447
Removing WinRPM dependency as trying with binary files
