### 2019-06-16 511ff67
pdDocHasSignature method added with testcases.
