+++ src/PDFIO.jl
@@ -24,6 +24,7 @@ export  PDDoc,
             pdDocGetPageRange,
             pdDocGetPageLabel,
             pdDocGetOutline,
+            pdDocHasSignature,
             pdDocValidateSignatures,
         PDPage,
             pdPageGetContents,
