### 2019-06-16 879f23f
Improving coverage and better handling of cert and key comparisons.
