### 2019-06-17 8f27726
Bumping up the version and ensuring .gitignore is updated.
