### 2019-06-17 6e38bf3
Digital Signature support (#63)
Consolidated check-ins for support for Digital Certificates.
