+++ src/PDCrypt.jl
@@ -1,0 +1,802 @@
+@static isfile(joinpath(dirname(@__FILE__), "..", "deps", "deps.jl")) ||
+        error("PDFIO not properly installed. Please run Pkg.build(\"PDFIO\")")
+
+include("../deps/deps.jl")
+
+using Dates
+
+const V_ASN1_UNDEF                    = Cint(-1)
+const V_ASN1_EOC                      = Cint(0)
+const V_ASN1_BOOLEAN                  = Cint(1)
+const V_ASN1_INTEGER                  = Cint(2)
+const V_ASN1_BIT_STRING               = Cint(3)
+const V_ASN1_OCTET_STRING             = Cint(4)
+const V_ASN1_NULL                     = Cint(5)
+const V_ASN1_OBJECT                   = Cint(6)
+const V_ASN1_OBJECT_DESCRIPTOR        = Cint(7)
+const V_ASN1_EXTERNAL                 = Cint(8)
+const V_ASN1_REAL                     = Cint(9)
+const V_ASN1_ENUMERATED               = Cint(10)
+const V_ASN1_UTF8STRING               = Cint(12)
+const V_ASN1_SEQUENCE                 = Cint(16)
+const V_ASN1_SET                      = Cint(17)
+const V_ASN1_NUMERICSTRING            = Cint(18)
+const V_ASN1_PRINTABLESTRING          = Cint(19)
+const V_ASN1_T61STRING                = Cint(20)
+const V_ASN1_TELETEXSTRING            = Cint(20)
+const V_ASN1_VIDEOTEXSTRING           = Cint(21)
+const V_ASN1_IA5STRING                = Cint(22)
+const V_ASN1_UTCTIME                  = Cint(23)
+const V_ASN1_GENERALIZEDTIME          = Cint(24)
+const V_ASN1_GRAPHICSTRING            = Cint(25)
+const V_ASN1_ISO64STRING              = Cint(26)
+const V_ASN1_VISIBLESTRING            = Cint(26)
+const V_ASN1_GENERALSTRING            = Cint(27)
+const V_ASN1_UNIVERSALSTRING          = Cint(28)
+const V_ASN1_BMPSTRING                = Cint(30)
+
+
+# Abstraction around OpenSSL libcrypto methods
+
+# Memory BIO from OpenSSL BIO interface as a Julia IO
+mutable struct BIO <: IO
+    data::Ptr{Cvoid}
+    function BIO(bio::Ptr{Cvoid})
+        this = new(bio)
+        finalizer(x->ccall((:BIO_vfree, libcrypto), Cvoid,
+                           (Ptr{Cvoid}, ), x.data), this)
+    end
+end
+
+# Readonly memory BIO
+function BIO(membuf::Vector{UInt8})
+    bio = ccall((:BIO_new_mem_buf, libcrypto), Ptr{Cvoid},
+                (Ptr{Cvoid}, Cint), pointer(membuf), length(membuf))
+    bio == C_NULL && openssl_error(0)
+    return BIO(bio)
+end
+
+# R/W Memory BIO
+function BIO()
+    buf_bio = ccall((:BIO_s_mem, libcrypto), Ptr{Cvoid}, ())
+    buf_bio == C_NULL && openssl_error(0)
+    bio = ccall((:BIO_new, libcrypto), Ptr{Cvoid}, (Ptr{Cvoid}, ), buf_bio)
+    bio == C_NULL && openssl_error(0)
+    return BIO(bio)    
+end
+
+function Base.read(bio::BIO)
[…]
