### 2019-06-19 11384cb
Updating README to represent quick getting started approach.
