+++ README.md
@@ -3,136 +3,126 @@
 [![*nix Status](https://travis-ci.org/sambitdash/PDFIO.jl.svg?branch=master)](https://travis-ci.org/sambitdash/PDFIO.jl)
 [![Win status](https://ci.appveyor.com/api/projects/status/9cocsctqdkx603q0?svg=true)](https://ci.appveyor.com/project/sambitdash/pdfio-jl)
 [![codecov.io](http://codecov.io/github/sambitdash/PDFIO.jl/coverage.svg?branch=master)](http://codecov.io/github/sambitdash/PDFIO.jl?branch=master)
-[![](https://img.shields.io/badge/docs-latest-blue.svg)](https://sambitdash.github.io/PDFIO.jl/dev)
-[![](https://img.shields.io/badge/docs-stable-blue.svg)](https://sambitdash.github.io/PDFIO.jl/stable)
+[![Doc Latest](https://img.shields.io/badge/docs-latest-blue.svg)](https://sambitdash.github.io/PDFIO.jl/dev)
+[![Doc Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://sambitdash.github.io/PDFIO.jl/stable)
 
-PDFIO is a simple PDF API focused on processing text from PDF files. The
-APIs is kept fairly simple and follows the PDF-32000-1:2008 specification as
-downloaded from
-[Adobe website](http://www.adobe.com/devnet/pdf/pdf_reference.html). The updated
-ISO versions can be downloaded from ISO websites. The API does not consider
-those versions yet. The API is developed in Julia keeping in mind the need for
-the file format parsing in native form. More over the object structure of PDF
-makes the API extend-able.
+PDFIO is a native Julia implementation for reading PDF files. It's an 100% Julia implementation of the PDF specification. Other than a few well established algorithms like flate decode (`zlib` library) or cryptographic operations (`openssl` library) almost all of the APIs are written in native Julia. 
 
-Technically,the PDFIO does not fully fit into the stream based file access APIs
-like [JuliaIO/FileIO](https://github.com/JuliaIO/FileIO.jl). However, there is a general interest
-to submit this API in that direction to ensure it's considered as a native file
-format understood by Julia.
+## Need for a PDF Reader API 
 
-## About PDF
+The following are some of the benefits of utilizing this approach:
 
-For people new to PDF development, PDF is a highly cross referenced file format
-without any specific linear location to read content. You can consider this as a
-form of a tree which needs to be read or understood as reading a tree. Hence, if
-your application demands you need not have to scan the full file format to get
-your job done. This is very different thinking than most file formats that are
-parsed from beginning to end.
+1. PDF files are in existence for over three decades. Implementations of the PDF writers are not always to the specification or they may even vary
+   significantly from vendor to vendor. Everytime, you get a new PDF
+   file there is a possibility that it may not work to the best interpretation
+   of the specification. A script based language makes it easier for the
+   consumer to quickly modify the code and enhance to their specific needs. 
+   
+2. When a higher level scripting language implements a C/C++ PDF library API the scope is confined to achieving certain high level application tasks like, graphics or text extraction; annotation or signature content extraction or page merging or extraction. However, this API represents the PDF specification as a model (in MVC parlance). Every object in PDF specification can be represented in some form through these APIs. Hence, objects can be utilized effectively to understand document structure or correlate documents in more meaningful ways. 
+   
+3. Potential to be extended as a PDF generator. Since, the API is written as an object model of PDF documents, it's easier to extend with additional PDF write or update capabilities.
+   
 
-The second part to remember is PDF is a derivative of a Page Description
-Language (PDL) like PostScript. The visual sanctity is considered most important
-than document structure. Although, the document structure was introduced in PDF
[…]
