### 2019-06-19 5bef094
Adding architectural details for PDF API.
