+++ docs/src/arch.md
@@ -1,0 +1,192 @@
+# Architecture and Design
+
+The design philosophy of PDF documents can be understood by the article written by John Warnock, founding CEO of Adobe Systems, called the [Camelot](http://www.planetpdf.com/planetpdf/pdfs/warnock_camelot.pdf) paper. Unlike, PostScript which is a Turing complete programming language, PDF is a file format. The operators in PDF do not leave a side effect on an execution stack, that can be further exploited by the subsequent operators. While one need not have full understanding of PostScript to understand PDF general complexities of a page description language like PostScript are seen in PDFs as well. Here are a few simple scenarios of similarities and dissimilarities of PDF vs PostScript. This will help you understand the design principles of PDF. 
+
+1. **Execution Order and Page**: PostScript is an execution engine. It would execute the code as the operators are received and results are left on the execution stack as side-effects. Depending on the device rendering capabilities,  the content PostScript would load the rendered output to the device on a call of `showpage`. PDF does not have the flexibility of execution of a programming language. Generally, PDF content are organized as page objects. The content to be rendered and the resources (fonts and images) are associated with a page. Moreover, the content is rendered in the order in which they appear. Unless otherwise specified, the ink is considered opaque. Thus a content rendered later may overwrite the content rendered earlier if there is an overlap.
+2. **Content Order vs. Reading Order**: PDLs focus on final look and feel of the output and not the reading order of the text or artifacts. For example, fonts tend to be an expensive resource to load. Once loaded an execution engine may like to render all the content utilizing the font resource earlier than render the artifacts in the reading order. A document optimized for such purpose will tend to organize all text utilizing a specific fonts at one place. 
+3. **Content Objects vs. Non-Content Objects**: Without getting much into the definition complexities, we can state a content leaves an artifact on a page and rendered. While some other objects may affect the user interface or provide additional document metadata information and they are not rendered on a PDF page. One example will be PDF outlines or navigation links. They only are shown on a user interface and not on a rendered page. 
+4. **Random vs. Sequential Access**: PDF files are designed for random access of the document. You can get to any page and access the contents of the page. While PostScript was not designed for that purpose. It typically will be loaded once and executed. 
+
+## PDF specification ISO-32000-1 and ISO-32000-2
+
+The API is developed around PDF specification, notably around ISO-32000-1, that can be downloaded from [Adobe](https://www.adobe.com/content/dam/acom/en/devnet/pdf/pdfs/PDF32000_2008.pdf). Some attempts have been made to consider certain features of PDF-2.0 ([ISO-32000-2](https://www.iso.org/standard/63534.html)). However, the attempts have been made with draft versions of the specification. The final version of the specification has not been reviewed for the implementations. Hence, the support is only experimental. 
+
+The API derives heavily from the PDF specification and tries to be as compliant to the specification as practical. Any significant discrepancies in the APIs have to be compared with the PDF specification for clarification and accuracy. The API is an extension to the PDF specification. Hence, it's recommended the users of the API  refer to the specification as well for a clearer understanding of the working of the API. Where a feature is best understood from the specification, reference to the API will be provided than elaborating it in the API documentation. 
+
+## PDF Components
+
+PDF specification identifies the following components of a file. For details about the components please refer to the Chapter-7 of the PDF specification 1.7. 
+
+1. Objects
[…]
