### 2019-06-23 d369970
Documentation update with minor code and test case updates.
