+++ paper/paper.bib
@@ -6,3 +6,12 @@
   url = {https://www.adobe.com/devnet/pdf/pdf_reference.html},
   note = {Provided by the Adobe Systems Incorporarted}
 }
+
+@Misc{Avik:2013,
+author={{Avik Sen}},
+year = 2013,
+title = {{Taro.jl - Read and write Excel, Word and PDF documents in Julia}}
+url = {https://github.com/aviks/Taro.jl}
+note = {In active development, last accessed 23rd June 2019}
+}
+
