+++ paper/paper.md
@@ -21,36 +21,121 @@ bibliography: paper.bib
 
 # Summary
 
-Portable Document Format (PDF) is the most ubiquitous file format today for text,
-scientific research, legal documentation and many other domains for information
-dissemination and presentation. A large body of text is currently available as
-archive in this format as well. Julia is an upcoming programming language in the
-field of data sciences with focus on text analysis. Extracting archived content 
-to text extracted form is highly beneficial to the language usage and adoption. 
+Portable Document Format (PDF) is the most ubiquitous file format for
+text, scientific research, legal documentation and many other domains
+for information dissemination and presentation. Being a final form
+format of choice, a large body of text is currently archived in this
+format. Julia is an upcoming programming language in the field of data
+sciences with focus on text analysis. Extracting archived content to
+text is highly beneficial to the language usage and adoption.
 
-``PDFIO`` is an API developed purely in Julia. Almost, all the functionality of 
-PDF understanding is entirely written from scratch in Julia with only exception 
-of usage of certain (de)compression codecs where standard open source softwares 
-are being used. The API is written keeping the low level understanding of the PDF
-specification. This enables these APIs to be used in real life data science 
-applications where higher level knowledge abstractions can be built over the low
-level libraries. However, an understanding of PDF specification(PDF-32000-1:2008)[@Adobe:2008]
-may be needed to accomplish certain advanced tasks with the APIs. 
+``PDFIO`` is an API developed purely in Julia. Almost, all the
+functionalities of PDF understanding is entirely written from scratch
+in Julia with only exception of certain (de)compression codecs and
+cryptography, where standard open source libraries are being used.
 
-The APIs are developed for text extraction as a major focus. The rendering and 
-printing aspects of PDF are not provided enough consideration. Secondly, a native 
-script based language was chosen as PDF specification is highly misunderstood 
-and interpretations significantly vary from document creator to creator. A script
-based language provides flexibility to fix issues as we discover more nuances in
-the new files we discover. Thirdly, every well-developed native library out in 
-the market need connectors and PDF being a standard should have native support in
-a modern language. Although, one can claim PDF is not the most ideal language for
-text processing, it's just ubiquitous and cannot be ignored.
+The following are some of the benefits of utilizing this approach:
+
+1. PDF files are in existence for over three decades. Implementations
+   of the PDF writers are not always accurate to the specification or
+   they may even vary significantly from vendor to vendor. Every time,
+   someone gets a new PDF file there is a possibility that it may not
+   work to the best interpretation of the specification. A script
+   based language makes it easier for the consumers to quickly modify
+   the code and enhance to their specific needs.
+   
+2. When a higher level scripting language implements a C/C++ PDF
+   library API, the scope is confined to achieving certain high level
+   application tasks like, graphics or text extraction; annotation or
+   signature content extraction or page merging or
+   extraction. However, this API represents the PDF specification as a
+   model (in Model, View and Controller parlance). Every object in PDF
+   specification can be represented in some form through these
+   APIs. Hence, objects can be utilized effectively to understand
+   document structure or correlate documents in more meaningful ways.
+    
+3. Potential to be extended as a PDF generator. Since, the API is
+   written as an object model of PDF documents, it's easier to extend
+   with additional PDF write or update capabilities.
+   
+There are also certain downsides to this approach:
+
+1. Any API that represents an object model of a document, tends to
[…]
