+++ src/CosDoc.jl
@@ -1,4 +1,3 @@
-<<<<<<< HEAD
 export
     CosDoc,
     cosDocOpen,
@@ -11,19 +10,6 @@ export
     merge_streams,
     find_ntree,
     readfrom
-=======
-export CosDoc,
-       cosDocOpen,
-       cosDocClose,
-       cosDocGetRoot,
-       cosDocGetInfo,
-       cosDocGetObject,
-       cosDocGetPageNumbers,
-       cosDocGetPageLabel,
-       merge_streams,
-       find_ntree,
-       readfrom
->>>>>>> 6e38bf397e4061932c335119c171acc5b7ed980c
 
 using Base: notnothing
 using ..Common
