### 2019-06-23 a19af96
Documentation revamp (#65)
