+++ docs/make.jl
@@ -8,11 +8,14 @@ using Documenter, PDFIO
 
 makedocs(
     format = Documenter.HTML(),
+    modules = [PDFIO],
     sitename = "PDFIO",
     pages = [
-        "intro.md",
-        "API Structure and Design" => "index.md",
-        "Digital Signatures" => "digsig.md"
+        "README.md",
+        "Architecture and Design" => "arch.md",
+        "Digital Signatures" => "digsig.md",
+        "API Reference" => "index.md",
+        "PDFIO License" => "LICENSE.md"
     ]
 )
 
