### 2019-06-23 51fc91a
Update paper.bib
Fixing errors.
