+++ paper/paper.bib
@@ -10,8 +10,8 @@
 @Misc{Avik:2013,
 author={{Avik Sen}},
 year = 2013,
-title = {{Taro.jl - Read and write Excel, Word and PDF documents in Julia}}
-url = {https://github.com/aviks/Taro.jl}
+title = {{Taro.jl - Read and write Excel, Word and PDF documents in Julia}},
+url = {https://github.com/aviks/Taro.jl},
 note = {In active development, last accessed 23rd June 2019}
 }
 
