### 2019-06-23 628664c
Update paper.md
fixing reference
