+++ v0.1.5/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"README/#PDFIO-1","page":"PDFIO","title":"PDFIO","text":"","category":"section"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"(Image: *nix Status) (Image: Win status) (Image: codecov.io) (Image: Doc Latest) (Image: Doc Stable)","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"PDFIO is a native Julia implementation for reading PDF files. It's an 100% Julia  implementation of the PDF specification. Other than a few well established  algorithms like flate decode (zlib library) or cryptographic operations  (openssl library) almost all of the APIs are written in native Julia. ","category":"page"},{"location":"README/#Need-for-a-PDF-Reader-API-1","page":"PDFIO","title":"Need for a PDF Reader API","text":"","category":"section"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"The following are some of the benefits of utilizing this approach:","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"PDF files are in existence for over three decades. Implementations of the PDF writers are not always to the specification or they may even vary significantly from vendor to vendor. Everytime, you get a new PDF file there is a possibility that it may not work to the best interpretation of the specification. A script based language makes it easier for the consumers to quickly modify the code and enhance to their specific needs. \nWhen a higher level scripting language implements a C/C++ PDF library API, the scope is confined to achieving certain high level application tasks like,  graphics or text extraction; annotation or signature content extraction or  page merging or extraction. However, this API represents the PDF  specification as a model (in MVC parlance). Every object in PDF  specification can be represented in some form through these APIs. Hence,  objects can be utilized effectively to understand document structure or  correlate documents in more meaningful ways. \nPotential to be extended as a PDF generator. Since, the API is written as an  object model of PDF documents, it's easier to extend with additional PDF  write or update capabilities.","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"There are also certain downsides to this approach:","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"Any API that represents an object model of a document, tends to carry the  complexity of introducing abstract objects, often opaque objects (handles)  that are merely representational for an API user. They may not have any  functional meaning. The methods tend to be granular than a method that can  complete a user level task. \nThe user may need to refer to the PDF specification for having a complete  semantic understanding.\nThe amount of code needed to carry out certain tasks can be substantially  higher. ","category":"page"},{"location":"README/#Illustration-1","page":"PDFIO","title":"Illustration","text":"","category":"section"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"A popular package Taro.jl that utilizes Java based Apache Tika, Apache POI and Apache FOP libraries for reading PDF and other file types may need the following code to extract text and other metadata from the document.","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"using Taro\nTaro.init()\nmeta, txtdata = Taro.extract(\"sample.pdf\");\n","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"While the same with PDFIO may look like below:","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"function getPDFText(src, out)\n    doc = pdDocOpen(src)\n    docinfo = pdDocGetInfo(doc)\n    open(out, \"w\") do io\n\t\tnpage = pdDocGetPageCount(doc)\n        for i=1:npage\n            page = pdDocG
[…]
