### 2019-07-05 7a0a98b
Implementation of PDF security handler for the following cases:
1. Standard security handler
2. R 2-5 versions only. R 6 has issues.
3. Streams with crypto filters need proper implementation.
4. Many minor coding or style issues observed
