+++ src/Cos.jl
@@ -13,5 +13,7 @@ include("CosStream.jl")
 include("CosReader.jl")
 include("CosObjStream.jl")
 include("CosDoc.jl")
+include("CosCrypt.jl")
+include("StdSecHandler.jl")
 
 end
