+++ src/Common.jl
@@ -5,5 +5,6 @@ include("errors.jl")
 include("Utils.jl")
 include("BufferParser.jl")
 include("CDObject.jl")
+include("LibCrypto.jl")
 
 end

[…]
