### 2019-07-07 e077020
Improving SecretBuffer usage.
