### 2019-07-08 a807f6a
Adding the new files for crypto and standard security handler. The following are now supported:
1. Support for AES-256 R6

TBD:

1. Support for streams which have crypt filter
2. Annoying warning messages on SecretBuffers on errors.
