+++ src/StdSecHandler.jl
@@ -1,0 +1,512 @@
+using Base: Unicode, SecretBuffer, SecretBuffer!, getpass, shred!
+
+const PASSWD_PADDING =
+    [0x28, 0xBF, 0x4E, 0x5E, 0x4E, 0x75, 0x8A, 0x41,
+     0x64, 0x00, 0x4E, 0x56, 0xFF, 0xFA, 0x01, 0x08,
+     0x2E, 0x2E, 0x00, 0xB6, 0xD0, 0x68, 0x3E, 0x80,
+     0x2F, 0x0C, 0xA9, 0xFE, 0x64, 0x53, 0x69, 0x7A]
+
+const AES_SUFFIX = [0x73, 0x41, 0x6C, 0x54] # sAlT
+
+function UTF8ToPDFEncoding(s::SecretBuffer)
+    r = SecretBuffer()
+    while !eof(s)
+        write(r, UnicodeToPDFEncoding(read(s, Char)))
+    end
+    return seekstart(r)
+end
+
+struct StdSecHandler <: SecHandler
+    v::Int
+    version::Tuple{Int, Int}
+    length::Int
+    r::Int
+    o::Vector{UInt8}
+    u::Vector{UInt8}
+    oe::Vector{UInt8}
+    ue::Vector{UInt8}
+    p::UInt32 
+    perms::Vector{UInt8}
+    encMetadata::Bool
+    id::Vector{UInt8}
+    cf::Union{CosDict, CosNullType}
+    stmf::CosName
+    strf::CosName
+    eff::CosName
+    password::Function
+    keys::Dict{CosName, Tuple{UInt32, Vector{UInt8}}}
+end
+
+const STORE_KEY = crypto_random(32)
+const STORE_IV  = crypto_random(16)
+
+function store_key(s::StdSecHandler, cfn::CosName,
+                   data::Tuple{UInt32, SecretBuffer})
+
+    skey = SecretBuffer(); iv = SecretBuffer()
+    write(skey, STORE_KEY); write(iv, STORE_IV)
+    cctx = CipherContext("aes_256_cbc", skey, iv, true)
+    shred!(skey); shred!(iv)
+
+    perm, key = data
+    c = update!(cctx, key)
+    append!(c, close(cctx))
+    s.keys[cfn] = (perm, c)
+    return data
+end
+
+function get_key(s::StdSecHandler, cfn::CosName)
+    permkey = get(s.keys, cfn, nothing)
+    permkey === nothing && return nothing
+    
+    skey = SecretBuffer(); iv = SecretBuffer()
+    write(skey, STORE_KEY); write(iv, STORE_IV)
+    cctx = CipherContext("aes_256_cbc", skey, iv, false)
+    shred!(skey); shred!(iv)
+    
+    perm, c = permkey
+    b = update!(cctx, c)
+    append!(b, close(cctx))
+    return perm, SecretBuffer!(b)
+end
+
+CryptParams(h::StdSecHandler, oi::CosIndirectObject) = 
+    CryptParams(h, oi.num, oi.gen, oi.obj)
+
+CryptParams(h::StdSecHandler, num::Int, gen::Int, o::CosObject) =
+    h.r < 4 ? CryptParams(num, gen, cn"StdCF") : error(E_INVALID_OBJECT)
+
+CryptParams(h::StdSecHandler, num::Int, gen::Int, o::CosString) =
+    CryptParams(num, gen, h.strf)
+
+CryptParams(h::StdSecHandler, num::Int, gen::Int, o::CosObjectStream) =
+    CryptParams(h, num, gen, o.stm)
+
+function CryptParams(h::StdSecHandler, num::Int, gen::Int, o::CosStream)
+    cfn = nothing
+    filters = get(o, cn"FFilter")
[…]
