### 2019-07-08 b8a3d9f
Handling of secured streams and crypt filters. Some special handling of CosObjStreams.
