### 2019-07-08 b5309a7
PDF page content may not be delimited properly.
1. In such cases look for the longest operator. Once found only that
portion has to be read by the parser.
2. While in case of CMaps whatever data is read is consumed. CMap reading
is very much limited to a very few sets of keywords. Hence, this
difference in handling.

qQq is a valid sequence of gsave, grestore in content stream.
