### 2019-07-09 5103388
Added test cases for file encryption
