+++ src/StdSecHandler.jl
@@ -8,12 +8,16 @@ const PASSWD_PADDING =
 
 const AES_SUFFIX = [0x73, 0x41, 0x6C, 0x54] # sAlT
 
-function UTF8ToPDFEncoding(s::SecretBuffer)
-    r = SecretBuffer()
-    while !eof(s)
-        write(r, UnicodeToPDFEncoding(read(s, Char)))
+function UTF8ToPDFEncoding!(s::SecretBuffer)
+    try
+        r = SecretBuffer()
+        while !eof(s)
+            write(r, UnicodeToPDFEncoding(read(s, Char)))
+        end
+        return seekstart(r)
+    finally
+        shred!(s)
     end
-    return seekstart(r)
 end
 
 struct StdSecHandler <: SecHandler
@@ -34,17 +38,15 @@ struct StdSecHandler <: SecHandler
     strf::CosName
     eff::CosName
     password::Function
+    skey_path::String
+    iv_path::String
     keys::Dict{CosName, Tuple{UInt32, Vector{UInt8}}}
 end
 
-const STORE_KEY = crypto_random(32)
-const STORE_IV  = crypto_random(16)
-
 function store_key(s::StdSecHandler, cfn::CosName,
                    data::Tuple{UInt32, SecretBuffer})
-
-    skey = SecretBuffer(); iv = SecretBuffer()
-    write(skey, STORE_KEY); write(iv, STORE_IV)
+    skey = SecretBuffer!(read(s.skey_path, 32))
+    iv   = SecretBuffer!(read(s.iv_path, 16))
     cctx = CipherContext("aes_256_cbc", skey, iv, true)
     shred!(skey); shred!(iv)
 
@@ -59,8 +61,8 @@ function get_key(s::StdSecHandler, cfn::CosName)
     permkey = get(s.keys, cfn, nothing)
     permkey === nothing && return nothing
     
-    skey = SecretBuffer(); iv = SecretBuffer()
-    write(skey, STORE_KEY); write(iv, STORE_IV)
+    skey = SecretBuffer!(read(s.skey_path, 32))
+    iv   = SecretBuffer!(read(s.iv_path, 16))
     cctx = CipherContext("aes_256_cbc", skey, iv, false)
     shred!(skey); shred!(iv)
     
@@ -131,8 +133,8 @@ function algo01(h::StdSecHandler, params::CryptParams,
     genarr = genarr[1:2]
     perm, key = get_key(h, params)
     n = div(h.length, 8)
-    n != key.size && error("Invalid encryption key length")
     md = shred!(key) do kc
+        n != kc.size && error("Invalid encryption key length")
         seekend(kc); write(kc, numarr); write(kc, genarr)
         !isRC4 && write(kc, AES_SUFFIX)
         mdctx = DigestContext("md5")
@@ -143,23 +145,29 @@ function algo01(h::StdSecHandler, params::CryptParams,
     key = SecretBuffer!(md[1:l])
     iv = SecretBuffer!(isRC4 ?     UInt8[] :
                        isencrypt ? crypto_random(16) : data[1:16])
-    cctx = CipherContext(isRC4 ? "rc4" : "aes_128_cbc", key, iv, isencrypt)
-    shred!(key); shred!(iv)
-    d = (isRC4 || isencrypt) ? update!(cctx, data) :
-                               update!(cctx, (@view data[17:end]))
-    append!(d, close(cctx))
-    return d
-end
+    try
+        cctx = CipherContext(isRC4 ? "rc4" : "aes_128_cbc", key, iv, isencrypt)
+        d = (isRC4 || isencrypt) ? update!(cctx, data) :
+            update!(cctx, (@view data[17:end]))
+        append!(d, close(cctx))
+        return d
+    finally
[…]
