### 2019-07-09 992fa5d
Removing font files added for test. No need to compare the zero width template
