### 2019-07-09 543f46a
Support for encrypted PDF files with standard crypto handler (#67)
* Implementation of PDF security handler for the following cases:

1. Standard security handler
2. R 2-5 versions only. R 6 has issues.
3. Streams with crypto filters need proper implementation.
4. Many minor coding or style issues observed

* Improving SecretBuffer usage.

* Adding the new files for crypto and standard security handler. The following are
now supported:

1. Support for AES-256 R6

TBD:

1. Support for streams which have crypt filter
2. Annoying warning messages on SecretBuffers on errors.

* Handling of secured
[…]
