+++ src/PDFIO.jl
@@ -60,6 +60,7 @@ export  CosDoc,
             cosDocClose,
             cosDocGetRoot,
             cosDocGetObject,
+            cosDocIsEncrypted,
         CosObject,
             CosNull, CosDict, CosString, CosArray, CosStream,
             CosIndirectObjectRef, set!,
