### 2019-07-10 123afb0
Update LICENSE.md
Updated the OpenSSL license text.
