### 2019-07-12 7a0ed21
PKI Security Handler for decrypting certificate based PDF files.
