### 2019-07-12 6bb4075
File may get locked in windows so forcing file closure.
