### 2019-07-12 695b9a0
Read the p12 file to a buffer. As long as the password is shredded it's not as severe a security flaw.
