+++ src/LibCrypto.jl
@@ -871,10 +871,10 @@ function get_rsa_digest(si::PKCS1SignedInfo, cert::Cert)
 end
 
 function read_pkcs12(fn::AbstractString, pw::SecretBuffer)
-    bio = BIO(String(fn))
-    p12 = ccall((:d2i_PKCS12_bio, libcrypto), Ptr{Cvoid},
-                (Ptr{Cvoid}, Ptr{Cvoid}), bio.data, C_NULL)
-    finalize(bio)
+    data = read(fn)
+    p12 = ccall((:d2i_PKCS12, libcrypto), Ptr{Cvoid},
+                (Ptr{Cvoid}, Ptr{Ptr{Cuchar}}, Clong),
+                C_NULL, Ref(pointer(data)), length(data))
     p12 == C_NULL && error("Unable to read $fn")
 
     xkey, xcert, xca = Ref(C_NULL), Ref(C_NULL), Ref(C_NULL)
