+++ src/StdSecHandler.jl
@@ -37,73 +37,10 @@ struct StdSecHandler <: SecHandler
     stmf::CosName
     strf::CosName
     eff::CosName
-    password::Function
+    access::Function
     skey_path::String
     iv_path::String
     keys::Dict{CosName, Tuple{UInt32, Vector{UInt8}}}
-end
-
-function store_key(s::StdSecHandler, cfn::CosName,
-                   data::Tuple{UInt32, SecretBuffer})
-    skey = SecretBuffer!(read(s.skey_path, 32))
-    iv   = SecretBuffer!(read(s.iv_path, 16))
-    cctx = CipherContext("aes_256_cbc", skey, iv, true)
-    shred!(skey); shred!(iv)
-
-    perm, key = data
-    c = update!(cctx, key)
-    append!(c, close(cctx))
-    s.keys[cfn] = (perm, c)
-    return data
-end
-
-function get_key(s::StdSecHandler, cfn::CosName)
-    permkey = get(s.keys, cfn, nothing)
-    permkey === nothing && return nothing
-    
-    skey = SecretBuffer!(read(s.skey_path, 32))
-    iv   = SecretBuffer!(read(s.iv_path, 16))
-    cctx = CipherContext("aes_256_cbc", skey, iv, false)
-    shred!(skey); shred!(iv)
-    
-    perm, c = permkey
-    b = update!(cctx, c)
-    append!(b, close(cctx))
-    return perm, SecretBuffer!(b)
-end
-
-CryptParams(h::StdSecHandler, oi::CosIndirectObject) = 
-    CryptParams(h, oi.num, oi.gen, oi.obj)
-
-CryptParams(h::StdSecHandler, num::Int, gen::Int, o::CosObject) =
-    h.r < 4 ? CryptParams(num, gen, cn"StdCF") : error(E_INVALID_OBJECT)
-
-CryptParams(h::StdSecHandler, num::Int, gen::Int, o::CosString) =
-    CryptParams(num, gen, h.strf)
-
-CryptParams(h::StdSecHandler, num::Int, gen::Int, o::CosObjectStream) =
-    CryptParams(h, num, gen, o.stm)
-
-# For Crypt filter chain ensure the filters before the crypt filters are removed.
-# Crypt filter parameters are associated with the document and the CosStream as
-# such may not have access to such parameters. Hence, the crypt filter has to be
-# decrypted when the document information is available.
-function CryptParams(h::StdSecHandler, num::Int, gen::Int, o::CosStream)
-    cfn = nothing
-    filters = get(o, cn"FFilter")
-    filters === CosNull && return CryptParams(num, gen, h.stmf)
-    if filters isa CosName
-        if cn"Crypt" === filters
-            params = get(o, cn"FDecodeParms")
-            cfn = get(params, cn"Name")
-        end
-    else
-        i = findnext(x -> x === cn"Crypt", get(filters), 1)
-        params = get(o, cn"FDecodeParms")
-        cfn = get(params[i], cn"Name")
-        i > 1 && cosStreamRemoveFilters(stm, until=(i-1))
-    end
-    return CryptParams(num, gen, cfn !== nothing ? cfn : h.stmf)
 end
 
 # Output 32-bytes padded password
@@ -118,56 +55,6 @@ function get_padded_pw(password::SecretBuffer)
         end
     end
     return seekstart(up)
-end
-
-function algo01(h::StdSecHandler, params::CryptParams,
-                data::AbstractVector{UInt8}, isencrypt::Bool)
-    num, gen, cfn = params.num, params.gen, params.cfn
-    cfm = get_cfm(h, cfn)
-    isRC4 = cfm === cn"V2"
[…]
