+++ src/PKISecHandler.jl
@@ -1,0 +1,133 @@
+using Base: prompt, getpass, SecretBuffer, SecretBuffer!, shred!
+using ..Common
+
+mutable struct PKISecHandler <: SecHandler
+    subtype::CosName
+    version::Tuple{Int, Int}
+    length::Int
+    r::Int
+    o::Vector{UInt8}
+    u::Vector{UInt8}
+    oe::Vector{UInt8}
+    ue::Vector{UInt8}
+    p::UInt32 
+    perms::Vector{UInt8}
+    id::Vector{UInt8}
+    cf::Union{CosDict, CosNullType}
+    stmf::CosName
+    strf::CosName
+    eff::CosName
+    access::Function
+    skey_path::String
+    iv_path::String
+    keys::Dict{CosName, Tuple{UInt32, Vector{UInt8}}}
+end
+
+function PKISecHandler(doc::CosDoc, subtype::CosName, access::Function)
+    enc = doc.encrypt
+    length = get(get(enc, cn"Length", CosInt(128)))
+    vrsn   = doc.version
+    subtype in [cn"adbe.pkcs7.s3", cn"adbe.pkcs7.s4", cn"adbe.pkcs7.s5"] ||
+        error("Unsupported release of security handler")
+
+    r = subtype === cn"adbe.pkcs7.s3" ? 2 :
+        subtype === cn"adbe.pkcs7.s4" ? 3 : 4
+    o      = Vector{UInt8}(get(enc, cn"O")) 
+    u      = Vector{UInt8}(get(enc, cn"U"))
+    oe     = Vector{UInt8}(get(enc, cn"OE"))
+    ue     = Vector{UInt8}(get(enc, cn"UE"))
+    p      = reinterpret(UInt32, Int32(get(get(enc, cn"P", CosInt(0)))))
+    perms  = Vector{UInt8}(get(enc, cn"Perms"))
+    ids = cosDocGetID(doc)
+    id = Vector{UInt8}(ids[1])
+
+    cf   = get(enc, cn"CF", CosDict())
+    if r < 4
+        d = CosDict()
+        set!(cf, cn"DefaultCryptFilter", d)
+        set!(d, cn"Type", cn"CryptFilter")
+        set!(d, cn"CFM", cn"V2")
+        set!(d, cn"AuthEvent", cn"DocOpen")
+        set!(d, cn"Recipients", get(enc, cn"Recipients"))
+        set!(d, cn"EncryptMetadata", get(enc, cn"EncryptMetadata", CosTrue))
+        set!(d, cn"Length", get(enc, cn"Length", CosInt(128)))
+        stmf = get(enc, cn"StmF", cn"DefaultCryptFilter")
+        strf = get(enc, cn"StrF", cn"DefaultCryptFilter")
+    else
+        stmf = get(enc, cn"StmF", cn"Identity")
+        strf = get(enc, cn"StrF", cn"Identity")
+    end
+    eff  = get(enc, cn"EFF", stmf)
+    eff !== stmf &&
+        Base.@warn("Embedded file streams without Crypt filters may not decrypt")
+    access === identity && (access = get_digital_id)
+    skey_path, io = get_tempfilepath()
+    # Ensure the files have enough bytes in them.
+    shred!(SecretBuffer!(crypto_random(rand(1000:10000)))) do s
+        write(io, s)
+        close(io)
+    end
+    iv_path, io = get_tempfilepath()
+    # Ensure the files have enough bytes in them.
+    shred!(SecretBuffer!(crypto_random(rand(1000:10000)))) do s
+        write(io, s)
+        close(io)
+    end
+    keys = Dict(cn"Identity" => (0xffffffff, Vector{UInt8}()))
[…]
