### 2019-07-29 f287e72
Adding better fonts and encoding support.
