+++ src/PDFonts.jl
@@ -1,13 +1,13 @@
 import ..Cos: CosXString
 
-export pdFontIsBold,
+export
+    pdFontIsBold,
     pdFontIsItalic,
     pdFontIsFixedW,
     pdFontIsAllCap,
     pdFontIsSmallCap
 
 using Rectangle
-
 
 #=
 Sample CMaps are available now as 8.cmap and 16.cmap in the test/files directory
@@ -44,23 +44,59 @@ function show(io::IO, cmap::CMap)
     show(io, cmap.range_map)
 end
 
+
+const FontUnicodeMapping = Union{Dict{UInt8, Char}, CMap, Nothing}
+
+#=
 mutable struct FontUnicodeMapping
     encoding::Dict{UInt8, Char}
     cmap::CMap
     hasCMap::Bool
     FontUnicodeMapping() = new(Dict{UInt8, Char}(), CMap(), false)
 end
+=#
 
-function merge_encoding!(fum::FontUnicodeMapping, encoding::CosName,
+function merge_encoding!(fum::Dict{UInt8, Char}, encoding::CosName,
                          doc::CosDoc, font::IDDRef{CosDict})
     encoding_mapping =
         encoding == cn"WinAnsiEncoding"   ? WINEncoding_to_Unicode :
         encoding == cn"MacRomanEncoding"  ? MACEncoding_to_Unicode :
         encoding == cn"MacExpertEncoding" ? MEXEncoding_to_Unicode :
                                             STDEncoding_to_Unicode
-    merge!(fum.encoding, encoding_mapping)
+    merge!(fum, encoding_mapping)
     return fum
 end
+
+abstract type FontType end
+struct FontType1    <: FontType end
+struct FontType3    <: FontType end
+struct FontMMType1  <: FontType end
+struct FontTrueType <: FontType end
+struct FontDefType  <: FontType end
+
+function FontType(subtype::CosName)
+    subtype === cn"Type1"      && return FontType1()
+    subtype === cn"Type3"      && return FontType3()
+    subtype === cn"TrueType"   && return FontTrueType()
+    subtype === cn"MMType1"    && return FontMMType1()
+    return FontDefType()
+end
+
+merge_encoding!(fum::FontUnicodeMapping, ftype::FontType,
+                doc::CosDoc, font::IDDRef{CosDict}) = fum
+
+function merge_encoding!(fum::Dict{UInt8, Char},
+                         ftype::Union{FontType1, FontMMType1},
+                         doc::CosDoc, font::IDDRef{CosDict})
+    basefont = cosDocGetObject(doc, font, cn"BaseFont")
+    basefont_with_subset = CDTextString(basefont)
+    basefont_str = rsplit(basefont_with_subset, '+';limit=2)[end]
+    enc = basefont_str == "Symbol"       ? SYMEncoding_to_Unicode :
+          basefont_str == "ZapfDingbats" ? ZAPEncoding_to_Unicode :
+                                           STDEncoding_to_Unicode
+    merge!(fum, enc)
+    return fum
+end    
 
 # for type 0 use cmap.
 # for symbol and zapfdingbats - use font encoding
@@ -68,26 +104,21 @@ end
 # Reading encoding from the font files in case of Symbolic fonts are not
 # supported.
 # Font subset is addressed with font name identification.
-function merge_encoding!(fum::FontUnicodeMapping, encoding::CosNullType,
-                        doc::CosDoc, font::IDDRef{CosDict})
+function merge_encoding!(fum::Dict{UInt8, Char}, encoding::CosNullType,
+                         doc::CosDoc, font::IDDRef{CosDict})
[…]
