### 2019-07-29 153ed2a
Removing commented away code.
