+++ src/PDPage.jl
@@ -268,16 +268,6 @@ end
     return load_page_objects(page, stm)
 end
 
-#=
-function populate_font_encoding(page, font, fontname)
-    if get(page.fums, fontname, CosNull) === CosNull
-        fum = FontUnicodeMapping()
-        merge_encoding!(fum, page.doc.cosDoc, font)
-        page.fums[fontname] = fum
-    end
-end
-=#
-
 function find_resource(page::PDPageImpl,
                        restype::CosName,
                        fontname::Union{CosName, CosNullType})
