### 2019-08-02 7c4c103
Enhancing the font unicode mapping framework.  (#71)
* Adding better fonts and encoding support.
