+++ src/PDPage.jl
@@ -7,11 +7,10 @@ export PDPage,
     pdPageGetMediaBox,
     pdPageGetCropBox,
     pdPageExtractText,
-    pdPageGetPageNumber
+    pdPageGetPageNumber,
+    pdPageEvalContent
 
 using ..Cos
-
-abstract type PDPage end
 
 """
 ```
@@ -270,14 +269,6 @@ end
     return load_page_objects(page, stm)
 end
 
-function populate_font_encoding(page, font, fontname)
-    if get(page.fums, fontname, CosNull) === CosNull
-        fum = FontUnicodeMapping()
-        merge_encoding!(fum, page.doc.cosDoc, font)
-        page.fums[fontname] = fum
-    end
-end
-
 function find_resource(page::PDPageImpl,
                        restype::CosName,
                        fontname::Union{CosName, CosNullType})
@@ -344,3 +335,4 @@ get_encoded_string(s::CosString, fontname::CosNullType, page::PDPage) =
 
 get_encoded_string(s::CosString, fontname::CosName, page::PDPage) =
     get_encoded_string(s, get(page.fonts, fontname, nothing))
+
