### 2019-08-03 09a017b
A basefont as standard 14 does not mean there cannot be additional encoding info.
