+++ src/PDFonts.jl
@@ -159,8 +159,8 @@ end
 function update_glyph_id_std_14(cosdoc, cosfont,
                                 glyph_name_to_cid, cid_to_glyph_name)
     basefont = cosDocGetObject(cosdoc, cosfont, cn"BaseFont")
-    basefont === CosNull && return false
-    String(basefont) in ADOBE_STD_14 || return false
+    basefont === CosNull && return 
+    String(basefont) in ADOBE_STD_14 || return
     gn2cid, cid2gn =
         basefont === cn"Symbol" ?
         (GlyphName_to_SYMEncoding, SYMEncoding_to_GlyphName) :
@@ -169,7 +169,7 @@ function update_glyph_id_std_14(cosdoc, cosfont,
         (GlyphName_to_STDEncoding, STDEncoding_to_GlyphName)
     merge!(glyph_name_to_cid, gn2cid)
     merge!(cid_to_glyph_name, cid2gn)
-    return true
+    return
 end
 
 function get_glyph_id_mapping(cosdoc::CosDoc, cosfont::IDD{CosDict})
@@ -179,10 +179,7 @@ function get_glyph_id_mapping(cosdoc::CosDoc, cosfont::IDD{CosDict})
     subtype = cosDocGetObject(cosdoc, cosfont, cn"Subtype")
     subtype === cn"Type0" && return glyph_name_to_cid, cid_to_glyph_name
 
-    update_glyph_id_std_14(cosdoc, cosfont,
-                           glyph_name_to_cid,
-                           cid_to_glyph_name) &&
-        return glyph_name_to_cid, cid_to_glyph_name
+    update_glyph_id_std_14(cosdoc, cosfont, glyph_name_to_cid, cid_to_glyph_name)
 
     encoding = cosDocGetObject(cosdoc, cosfont, cn"Encoding")
     encoding === CosNull && return glyph_name_to_cid, cid_to_glyph_name
