### 2019-08-11 b7e0c11
Update .travis.yml
Adding windows build
