### 2019-09-01 #1453 @alexhanna
Ee, sorry about that @labarba. I haven't heard back from @jameshclrk. I emailed him on June 16. Should we remove him from the review and attempt to find another reviewer?

@malmaud -- can you have another pass at this package? Do the documentation changes address your concerns?
