### 2019-09-13 0711083
Update LICENSE.md
Added licensing advertising text.
