### 2019-09-13 60033fc
Update README.md
OpenSSL licensing information in main documentation.
