### 2019-10-05 b9cba45
Ensuring the unicode mapping as well as the encoding to glyph mapping use similar code paths for Type1 fonts.
