### 2019-10-13 #1453 @alexhanna
:wave: @malmaud are you able to continue this review?
Removing @jameshclrk from this review, given non-response.
@vchuravy, @ahwillia, @ChrisRackauckas -- are any of you available to review this?
