### 2019-10-16 #1453 @malmaud
OK, tested it out. Functionality seems to work - I was able to load a PDF of my own and extract the its text and query the metadata, and it all seemed accurate. I'm not a particular domain expert on PDF reader tools but it least seems to do what it promises to do in its README.

I see examples have been added to most of the docstrings of the exported functions and an example is given on the README, which is great, as well as a welcome addition of a table of contents which organized the API. I still find that many of these docstrings have some spelling or grammatical errors, but it's good enough to figure out how to use the package.

I think the corresponding JOSS paper still needs a bit of work, though - there are quite a few remaining grammar and punctuation issues which make it difficult to read. I also find it pretty hard to understand the claims being made. I don't understand the comments about MVC architecture in benefit number 2, and benefit 3 talks about the ability to modify PDFs, although the package doesn't actually support that at the moment. I don't understand in downside 1 what it means for a handle to be 'merely representational'.

I think if the paper can be made more clear and some of the more nebulous claims and phrases rectified, this should be in good shape though.
