### 2019-10-17 be4f9b3
Update to the JOSS paper based on the comments.
