+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'PDFIO: PDF Reader Library for native Julia'
+title: PDFIO: PDF Reader Library for native Julia'
 tags:
   - Julia
   - PDF
@@ -22,54 +22,81 @@ bibliography: paper.bib
 # Summary
 
 Portable Document Format (PDF) is the most ubiquitous file format for
-text, scientific research, legal documentation and many other domains
-for information dissemination and presentation. Being a final form
+text, scientific research, legal documentation and many other fields
+for information presentation and dissemination. Being a final form
 format of choice, a large body of text is currently archived in this
 format. Julia is an upcoming programming language in the field of data
-sciences with focus on text analysis. Extracting archived content to
-text is highly beneficial to the language usage and adoption.
+sciences. Extracting archived content and understanding document 
+metadata is beneficial to the language usage.
 
-``PDFIO`` is an API developed purely in Julia. Almost, all the
+`PDFIO` is an API developed purely in Julia. Almost, all the
 functionalities of PDF understanding is entirely written from scratch
-in Julia with only exception of certain (de)compression codecs and
+in Julia with the only exception of certain (de)compression codecs and
 cryptography, where standard open source libraries are being used.
 
 The following are some of the benefits of utilizing this approach:
 
 1. PDF files are in existence for over three decades. Implementations
-   of the PDF writers are not always accurate to the specification or
-   they may even vary significantly from vendor to vendor. Every time,
-   someone gets a new PDF file there is a possibility that it may not
-   work to the best interpretation of the specification. A script
-   based language makes it easier for the consumers to quickly modify
-   the code and enhance to their specific needs.
+   of the PDF writers are not always accurate to the specification.
+   They may even vary significantly from vendor to vendor. Every time 
+   someone gets a new PDF file, there is a possibility that file
+   may not be interpreted as per the specification. A script based 
+   language makes it easier for the consumers to quickly modify and 
+   enhance the code to their specific needs.
    
 2. When a higher level scripting language implements a C/C++ PDF
-   library API, the scope is confined to achieving certain high level
-   application tasks like, graphics or text extraction; annotation or
-   signature content extraction or page merging or
-   extraction. However, this API represents the PDF specification as a
-   model (in Model, View and Controller parlance). Every object in PDF
-   specification can be represented in some form through these
-   APIs. Hence, objects can be utilized effectively to understand
-   document structure or correlate documents in more meaningful ways.
+   library API, the scope is kept limited to achieving certain high 
+   level tasks like, graphics or text extraction; annotation or
+   signature content extraction; or page extraction or merging. 
+   
+   However, `PDFIO` represents the PDF specification as a model in the 
+   Model, View and Controller parlance. A PDF file can be represented 
+   as a collection of interconnected Julia structures. Those 
+   structures can be utilized in granular tasks or simply can be used 
+   to understand the structure of the PDF document. 
+
+   As per the PDF specification, text can be presented as part of the
+   page content stream or inside PDF page annotations. An API like 
+   `PDFIO` can create two categories of object types. One representing
+   the text object inside the content stream and the other for the 
+   text inside an annotation object. Thus, providing flexibility to 
+   the API user. 
     
-3. Potential to be extended as a PDF generator. Since, the API is
-   written as an object model of PDF documents, it's easier to extend
-   with additional PDF write or update capabilities.
+3. Since, the API is written as an object model of PDF documents, it's 
+   easier to extend with additional PDF write or update capabilities. 
+   Although, the current implementation does not provide the PDF 
+   writing capabilities, the foundation has been laid for future 
+   extension.
    
[…]
