### 2019-10-17 #1453 @sambitdash
Thanks @malmaud for your review.

I have modified the paper based on your inputs.
