### 2019-11-01 #1453 @jarvist
I've just been reviewing this software + paper, but I can't seem to edit the checkboxes (also, they are still allocated to James Clark. @alexhanna - is this something you can fix?

Generally I'm happy with the package... it installs and tests run. I think the latest draft of the paper looks good.

However, I still think the package README.md could be edited down to make it more clear, and make it easier for a user to understand what the functionality is, and to get started.
For instance, you can:-
`julia> pdPageExtractText(stdout, pdDocGetPage(pdDocOpen("10.21105.joss.01453.pdf"), 1))`
from the REPL to extract the first page of the JOSS review into the terminal. This is pretty cool functionality, but I had to find this out by fiddling around with it.

Are there any other high level access functions than `pdPageExtractText` ?

From the point of view of the Package/API (not having read the source code), the namespace is quite confusing: the package is `PDFIO` but then all the exported functions are prepended with `pd` which seems superfluous for Julia, because everything is already under the `PDFIO` namespace of the module.
Similarly, the function arguments seem to vary quite a bit in placing and expectation. Does pdPageExtractText need an io object as the first argument? A default if single argument to return the text object would make a lot of sense. (Making the `stdout` in my example above superfluous.)
Also, the library seems to fail on Google-docs exported PDFs. (My first test case.) I will file an issue.
