### 2019-11-01 #1453 @alexhanna
You shouldn't need to edit it. You can click the checkboxes without editing. Does that work for you?
