### 2019-11-01 #1453 @jarvist
No joy there either. (I think the GUI checkbox clicking is predicated by edit permissions.)
