### 2019-11-01 f98d449
PDFs generated out of Google Docs create CID Widths with float unlike most cases where integer is used as widths.
