### 2019-11-01 #1453 @sambitdash
Hi @jarvist,

Thanks a lot for taking time to review the software and the paper. Also reporting the bug on the product.
>
> Are there any other high level access functions than `pdPageExtractText` ?

All methods starting with `pd` documented under: https://sambitdash.github.io/PDFIO.jl/dev/#PD-1 are high level functions.

>
> From the point of view of the Package/API (not having read the source code), the namespace is quite confusing: the package is `PDFIO` but then all the exported functions are prepended with `pd` which seems superfluous for Julia, because everything is already under the `PDFIO` namespace of the module.

`pd` does not stand for `PDFIO` but `PDF Document` or simply called `PD Layer` in most PDF format architectures. This is to differentiate from low level functions called `COS Layer` which requires you to understand the PDF objects like dictionary, arrays to be able to understand the PDF file. An explanation of these can be seen at: https://sambitdash.github.io/PDFIO.jl/dev/arch/

Moreover, the similar nomenclature is used by other established PDF libraries in the market. Hence, people who have used other PDF libraries may quickly understand the scope of the APIs.

> Similarly, the function arguments seem to vary quite a bit in placing and expectation. Does pdPageExtractText need an io object as the first argument? A default if single argument to return the text object would make a lot of sense. (Making the `stdout` in my example above superfluous.)

1. PDF files can be large expanding up to several megabytes and can technically grow to gigabytes. With extract text returning strings this may lead to several copies of large strings. IO provides a cleaner interface for the same. Many a times users may extract all pages and `IO` provides a possibility to append pages over one another.
2. PDF being an encrypted file format immutable extracted text should not be left in the memory which is a definite possibility with usage of strings. IO can be used to immediately write to a file or clear up the IOBuffer at the end of using the data.

> Also, the library seems to fail on Google-docs exported PDFs. (My first test case.) I will file an issue.

This has been fixed. If you think this is a gating requirement for the paper, please let me know so that I can release a new version of the software.
