+++ .travis.yml
@@ -1,7 +1,7 @@
 language: julia
 
 julia:
-  - 1.1
+  - 1.3
   - nightly
 
 os:
@@ -21,7 +21,7 @@ after_success:
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.1
+      julia: 1.3
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
