### 2019-11-02 f44bf15
Setting the version to 0.1.8
