### 2019-11-02 #1453 @jarvist
> This has been fixed. If you think this is a gating requirement for the paper, please let me know so that I can release a new version of the software.

I don't think this is a gating requirement.

However, I do think the README.md should be edited to make it more clear.

- Some kind of interactive example would really help: I had to read and process your function given in the illustration, to slowly understand enough to be able to play with it on the REPL.

- Merging the `illustration` and `Sample Code` sections might make sense. Most on the non-code in `Sample Code` I think would be clearer as in-line comments within the code.

- I don't think the references to `Taro.jl` are necessary in the main text, instead they should be part of a `Alternative Software` or similar section much further down.
