+++ .travis.yml
@@ -1,7 +1,8 @@
 language: julia
 
 julia:
-  - 1.1
+  - 1.0
+  - 1.2
   - nightly
 
 os:
