+++ appveyor.yml
@@ -1,6 +1,7 @@
 environment:
   matrix:
-  - julia_version: 1
+  - julia_version: 1.0
+  - julia_version: 1.2
   - julia_version: nightly
 
 platform:
