### 2019-11-03 5044a4d
Code refactoring and Setting Rectangle version to latest registered.
