+++ Project.toml
@@ -22,7 +22,7 @@ AbstractTrees = ">= 0.2.1"
 AdobeGlyphList = ">= 0.1.1"
 BinDeps = ">= 0.8.10"
 LabelNumerals = ">= 0.1.0"
-Rectangle = ">= 0.1.2"
+Rectangle = ">= 0.1.1"
 RomanNumerals = ">= 0.3.1"
 julia = "1.0"
 
