### 2019-11-03 da9fb76
Modified the README.md based on JOSS comments.
