+++ README.md
@@ -90,7 +90,7 @@ The above mentioned code takes a PDF file `src` as input and writes the text dat
 ```julia {.line_numbers}
 """
 ​```
-	getPDFText(src, out) -> Dict 
+    getPDFText(src, out) -> Dict 
 ​```
 - src - Input PDF file from where text is to be extracted
 - out - Output TXT file where the output will be written
@@ -99,27 +99,27 @@ return - A dictionary containing metadata of the document
 function getPDFText(src, out)
     # handle that can be used for subsequence operations on the document.
     doc = pdDocOpen(src)
-	
+    
     # Metadata extracted from the PDF document. 
-	# This value is retained and returned as the return from the function. 
+    # This value is retained and returned as the return from the function. 
     docinfo = pdDocGetInfo(doc) 
     open(out, "w") do io
-	
-		# Returns number of pages in the document		
-		npage = pdDocGetPageCount(doc)
+    
+        # Returns number of pages in the document       
+        npage = pdDocGetPageCount(doc)
 
         for i=1:npage
-		
+        
             # handle to the specific page given the number index. 
             page = pdDocGetPage(doc, i)
-			
-			# Extract text from the page and write it to the output file.
+            
+            # Extract text from the page and write it to the output file.
             pdPageExtractText(io, page)
 
-		end
+        end
     end
-	# Close the document handle. 
-	# The doc handle should not be used after this call
+    # Close the document handle. 
+    # The doc handle should not be used after this call
     pdDocClose(doc)
     return docinfo
 end
