+++ v0.1.8/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"README/#PDFIO-1","page":"PDFIO","title":"PDFIO","text":"","category":"section"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"(Image: *nix Status) (Image: Win status) (Image: codecov.io) (Image: Doc Latest) (Image: Doc Stable)","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"PDFIO is a native Julia implementation for reading PDF files. It's an 100% Julia  implementation of the PDF specification. Other than a few well established  algorithms like flate decode (zlib library) or cryptographic operations  (openssl library) almost all of the APIs are written in native Julia. ","category":"page"},{"location":"README/#Need-for-a-PDF-Reader-API-1","page":"PDFIO","title":"Need for a PDF Reader API","text":"","category":"section"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"The following are some of the benefits of utilizing this approach:","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"PDF files are in existence for over three decades. Implementations of the PDF writers are not always to the specification or they may even vary significantly from vendor to vendor. Everytime, you get a new PDF file there is a possibility that it may not work to the best interpretation of the specification. A script based language makes it easier for the consumers to quickly modify the code and enhance to their specific needs. \nWhen a higher level scripting language implements a C/C++ PDF library API, the scope is kept limited to achieving certain high  level tasks like, graphics or text extraction; annotation or signature content extraction; or page extraction or merging. \nHowever, PDFIO represents the PDF specification as a model in the  Model, View and Controller parlance. A PDF file can be represented  as a collection of interconnected Julia structures. Those  structures can be utilized in granular tasks or simply can be used  to understand the structure of the PDF document. \nAs per the PDF specification, text can be presented as part of the page content stream or inside PDF page annotations. An API like  PDFIO can create two categories of object types. One representing the text object inside the content stream and the other for the  text inside an annotation object. Thus, providing flexibility to  the API user. \nSince, the API is written as an object model of PDF documents, it's  easier to extend with additional PDF write or update capabilities.  Although, the current implementation does not provide the PDF  writing capabilities, the foundation has been laid for future  extension.","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"There are also certain downsides to this approach:","category":"page"},{"location":"README/#","page":"PDFIO","title":"PDFIO","text":"Any API that represents an object model of a document, tends to carry the complexity of introducing abstract objects. They can be opaque objects (handles) that are representational specific to the  API. They may not have any functional meaning. The methods are granular and may not complete one use level task. The amount of code needed to complete a user level task can be substantially higher. \nIn PDFIO the following steps have to be carried out:  a. Open the PDF document and obtain the document handle.   b. Query the document handle for all the pages in the document.  c. Iterate the pages and obtain the page object handles for each of    the pages.   d. Extract the text from the page objects and write to a file IO.   e. Close the document ensuring all the document resources are     reclaimed.\nThe API user may need to refer to the PDF specification (PDF-32000-1:2008)[@Adobe:2008] for semantic understanding of PDF  files in accomplishing some of the tasks. For example, the workflow  of PDF text extraction above is a natural extension from how text is  represented in a PDF file as per the specification. A PDF file is  composed of pages and text is represented inside each page content  object. The object model of PDFIO is a Julia language  representation of the PDF specification. ","category":"page"},{"location":"README/#Installation-1","page":"PDFIO","ti
[…]
