### 2019-11-03 642d320
Revert "Dropping 32-bit suport"
This reverts commit b2d5f9409d761a30296d480a583cb8c171bc0086.
