+++ appveyor.yml
@@ -4,6 +4,7 @@ environment:
   - julia_version: nightly
 
 platform:
+  - x86 # 32-bit
   - x64 # 64-bit
 
 # # Uncomment the following lines to allow failures on nightly julia
