+++ src/Inflate.jl
@@ -45,7 +45,10 @@ const Z_MEM_ERROR     = -4
 const Z_BUF_ERROR     = -5
 const Z_VERSION_ERROR = -6
 
-using Zlib_jll: libz
+@static isfile(joinpath(dirname(@__FILE__),"..","deps","deps.jl")) ||
+        error("PDFIO not properly installed. Please run Pkg.build(\"PDFIO\")")
+
+include("../deps/deps.jl")
 
 _zlibVersion() = ccall((:zlibVersion, libz), Ptr{Cstring}, ())
 
