### 2019-11-03 ecce789
Ensuring 1.0 support is kept intact as well.
