+++ src/Inflate.jl
@@ -45,10 +45,15 @@ const Z_MEM_ERROR     = -4
 const Z_BUF_ERROR     = -5
 const Z_VERSION_ERROR = -6
 
-@static isfile(joinpath(dirname(@__FILE__),"..","deps","deps.jl")) ||
+@static if Base.VERSION > v"1.3-" && !(Sys.iswindows() && Sys.WORD_SIZE == 32)
+    using Pkg
+    Pkg.add("Zlib_jll")
+    using Zlib_jll: libz
+else
+    isfile(joinpath(dirname(@__FILE__),"..","deps","deps.jl")) ||
         error("PDFIO not properly installed. Please run Pkg.build(\"PDFIO\")")
-
-include("../deps/deps.jl")
+    include("../deps/deps.jl")
+end
 
 _zlibVersion() = ccall((:zlibVersion, libz), Ptr{Cstring}, ())
 
