+++ appveyor.yml
@@ -1,18 +1,18 @@
 environment:
   matrix:
   - julia_version: 1.0
-  - julia_version: 1.2
+  - julia_version: 1.3
   - julia_version: nightly
 
 platform:
   - x86 # 32-bit
   - x64 # 64-bit
 
-# # Uncomment the following lines to allow failures on nightly julia
-# # (tests will run but not make your overall status red)
-# matrix:
-#   allow_failures:
-#   - julia_version: latest
+# Uncomment the following lines to allow failures on nightly julia
+# (tests will run but not make your overall status red)
+matrix:
+   allow_failures:
+   - julia_version: nightly
 
 branches:
   only:
