### 2019-11-14 #1453 @alexhanna
I've also added a few changes to the paper text in https://github.com/sambitdash/PDFIO.jl/pull/77.
