### 2019-11-14 #1453 @alexhanna
Thanks for merging those, @sambitdash.
@openjournals/joss-eics this one is good to go.
