### 2019-11-14 #1453 @labarba
The latest release in the repository is v0.1.8 — is that the reviewed version? Should we update the version number here?
