### 2019-11-14 #1453 @sambitdash
The version number can be updated to 0.1.8. The latest update to the README is incorporated in it.
