### 2019-11-14 #1453 @labarba
Many thanks to our editor: @alexhanna, and the reviewers: @malmaud, @jarvist — your contributions to JOSS are greatly appreciated 🙏
